% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_module.R
\docType{data}
\name{.__Extra__}
\alias{.__Extra__}
\title{Extra Earth Engine dictionary}
\format{
Earth Engine module
}
\usage{
.__Extra__
}
\description{
Interface to Extra Earth Engine module. Provides access to the top level
classes and functions as well as sub-modules (e.g. \code{ee$Image},
\code{ee$FeatureCollection$first}, etc.).
}
\examples{
\dontrun{
library(rgee)
library(rgeeExtra)

# ee_install_extra()
ee_Initialize()

ee_img1 <- ee$Image(0)
ee_img2 <- ee$Image(0)

ee_img3 <- ee_img1 * ee_img2 +  ee_img2
print(ee_img3)
}
}
\keyword{datasets}
