% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAMapclass.R
\docType{class}
\name{EarthEngineMap-class}
\alias{EarthEngineMap-class}
\title{Class EarthEngineMap}
\description{
Class EarthEngineMap
}
\section{Slots}{

\describe{
\item{\code{object}}{the spatial object}

\item{\code{map}}{the leaflet map object}
}}

