% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{ee.ImageCollection}
\alias{ee.ImageCollection}
\title{Soy una bonita funcion que no hace nada :)}
\usage{
ee.ImageCollection(ID)
}
\arguments{
\item{ID}{holi}
}
\description{
Soy una bonita funcion que no hace nada :)
}
