% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_module.R
\docType{data}
\name{ee}
\alias{ee}
\title{Main Earth Engine module}
\format{
Earth Engine module
}
\usage{
ee
}
\description{
Interface to main Earth Engine module. Provides access to top level classes
and functions as well as sub-modules (e.g. \code{ee$Image},
\code{ee$FeatureCollection$first}, etc.).
}
\examples{
\dontrun{
library(rgee)

ee_Initialize()

ee_img <- ee$Image(0)
ee_ic <- ee$ImageCollection(ee_img)

print(ee_img$getInfo())
print(ee_ic$getInfo())
}
}
\keyword{datasets}
