% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_geometry.R
\name{ee_as_eegeom}
\alias{ee_as_eegeom}
\title{Convert foreign geometry object to an Earth Engine object}
\usage{
ee_as_eegeom(x)
}
\arguments{
\item{x}{objet to be converted into an ee_Geometry.}
}
\description{
Convert foreign geometry object to an Earth Engine object
}
