% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_aux.R
\name{ee_as_geom}
\alias{ee_as_geom}
\title{Convert sf object to an ee_Geometry object}
\usage{
ee_as_geom(x)
}
\arguments{
\item{x}{objet to be converted into an ee_Geometry.}
}
\description{
Convert sf object to an ee_Geometry object
}
