% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map.R
\name{ee_as_mapview}
\alias{ee_as_mapview}
\title{Convert an EarthEngineMap object into a mapview object}
\usage{
ee_as_mapview(x)
}
\arguments{
\item{x}{An EarthEngineMap object.}
}
\description{
Convert an EarthEngineMap object into a mapview object
}
