% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_as_stars.R
\name{ee_as_stars}
\alias{ee_as_stars}
\title{Create a stars object based on a Earth Engine image}
\usage{
ee_as_stars(x, region, scale, palette = NULL, min = 0, max = 1,
  crs = "EPSG:4326")
}
\arguments{
\item{x}{TODO}

\item{region}{TODO}

\item{scale}{TODO}

\item{palette}{TODO}

\item{min}{TODO}

\item{max}{TODO}

\item{crs}{TODO}
}
\description{
Create a stars object based on a Earth Engine image
}
