% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_as_thumbnail.R
\name{ee_as_thumbnail}
\alias{ee_as_thumbnail}
\title{Create a stars object based on a Earth Engine thumbnail}
\usage{
ee_as_thumbnail(x, region, scale, vizparams = NULL, crs = "EPSG:4326")
}
\arguments{
\item{x}{Earth Engine Image object}

\item{region}{Geospatial region of the image (E,S,W,N, earth engine geometry object,
GeoJSON, or sfg). By default, the whole image will be used.}

\item{scale}{spatial pixel size.}

\item{vizparams}{list; visualization parameters. See Details.}

\item{crs}{The target projection e.g. 'EPSG:3857'. Defaults to WGS84 ('EPSG:4326').}
}
\description{
Create a stars object based on a Earth Engine thumbnail
}
\details{
The \code{ee_as_thumbnail} function shares parameters with \code{ee_map}. The parameters are:

\tabular{lll}{
\strong{Parameter}\tab \strong{Description}\tab\strong{Type}\cr
\strong{bands}     \tab  Comma-delimited list of three band names to be mapped to RGB                \tab  list                                               \cr
\strong{min}       \tab  Value(s) to map to 0                                                        \tab  number or list of three numbers, one for each band \cr
\strong{max}       \tab  Value(s) to map to 255                                                      \tab  number or list of three numbers, one for each band \cr
\strong{gain}      \tab  Value(s) by which to multiply each pixel value                              \tab  number or list of three numbers, one for each band \cr
\strong{bias}      \tab  Value(s) to add to each DN                                                  \tab  number or list of three numbers, one for each band \cr
\strong{gamma}     \tab  Gamma correction factor(s)                                                  \tab  number or list of three numbers, one for each band \cr
\strong{palette}  \tab  List of CSS-style color strings (single-band images only)                   \tab  comma-separated list of hex strings                \cr
\strong{opacity}   \tab  The opacity of the layer (0.0 is fully transparent and 1.0 is fully opaque) \tab  number                                             \cr
\strong{format}    \tab  Either "jpg" or "png"                                                       \tab  string \cr
}
}
\examples{
\dontrun{
library(rgee)
library(sf)
library(stars)
ee_Initialize()

# Fetch a digital elevation model.
image <- ee$Image('CGIAR/SRTM90_V4')

# params
palette <- c('00A600','63C600','E6E600','E9BD3A','ECB176','EFC2B3','F2F2F2')
vizparams1 <- list(min=0, max=3000)
vizparams2 <- list(min=0, max=3000, palette=palette)
outer = list(matrix(c(-84.6, 15.7,-84.6, -55.9,-32.9, -55.9,-32.9, 15.7,-84.6, 15.7),
                    ncol=2, byrow=TRUE))
region <- st_polygon(outer) \%>\% sf_as_ee
# plot1
my_thumbnail <- ee_as_thumbnail(x = image,region = region,scale = 1, vizparams = vizparams1)
plot(my_thumbnail)

# plot2
my_thumbnail <- ee_as_thumbnail(x = image,region = region,scale = 1,vizparams = vizparams2)
image(my_thumbnail, rgb = c(3,2,1))
}
}
