% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_check.R, R/ee_manage.R
\name{ee_check-tools}
\alias{ee_check-tools}
\alias{ee_check}
\alias{ee_check_python}
\alias{ee_check_rgee_python_packages}
\alias{ee_check_drivers}
\alias{ee_check_credentials}
\alias{ee_install_drivers}
\alias{ee_install_rgee_python_packages}
\alias{ee_install_python_package}
\alias{ee_install_python_ee}
\alias{ee_get_earthengine_path}
\alias{ee_remove_credentials}
\alias{ee_remove_drivers}
\alias{ee_manage_create}
\alias{ee_manage_delete}
\alias{ee_manage_assetlist}
\alias{ee_manage_quota}
\alias{ee_manage_size}
\alias{ee_manage_copy}
\alias{ee_manage_move}
\alias{ee_manage_set_properties}
\alias{ee_manage_delete_properties}
\alias{ee_manage_assets_access}
\title{Interface to handle rgee installation requirements}
\usage{
ee_check()

ee_check_python()

ee_check_rgee_python_packages()

ee_check_drivers(display_in_browser = TRUE)

ee_check_credentials()

ee_install_drivers(version)

ee_install_rgee_python_packages(conda = FALSE)

ee_install_python_package(pypackage, conda = FALSE)

ee_install_python_ee(conda = FALSE)

ee_get_earthengine_path()

ee_remove_credentials(quiet = TRUE)

ee_remove_drivers(quiet = TRUE)

ee_manage_create(path_asset, asset_type = "folder", quiet = FALSE)

ee_manage_delete(path_asset, quiet = FALSE)

ee_manage_assetlist(path_asset, quiet = FALSE)

ee_manage_quota()

ee_manage_size(path_asset)

ee_manage_copy(path_asset, final_path, quiet = FALSE)

ee_manage_move(path_asset, final_path, quiet = FALSE)

ee_manage_set_properties(path_asset, properties)

ee_manage_delete_properties(path_asset, property)

ee_manage_assets_access(path_asset,
  acl = getOption("rgee.manage.getAssetAcl"), quiet = FALSE)
}
\arguments{
\item{display_in_browser}{TODO}

\item{version}{TODO}

\item{conda}{TODO}

\item{pypackage}{TODO}

\item{quiet}{logical; suppress info message}

\item{path_asset}{a character vector containing a single path name.}

\item{asset_type}{a character vector containing the asset type. 'folder' or 'imagecollection'.}

\item{quiet}{logical; suppress info message.}
}
\description{
R functions for checking, installing and removing: drivers (geckodrivers), credentials and python packages.
}
\details{
It is neccessary restart RStudio to observe change when installing a python packages. rgee only is
compatible with Python 3.5 >.
}
\examples{
\dontrun{
library(rgee)

# Recommended way to install external dependencies of rgee for Linux users
virtualenv_remove("rgee")
virtualenv_create("rgee", python = "python3.5")
use_virtualenv("rgee")

ee_check()
ee_install_rgee_python_packages() # Install rgee python packages
ee_install_drivers() # Install selenim drivers
ee_get_credentials() # Get credentials
ee_check()

# Recommended way to install external dependencies of rgee for Windows users
}
}
