% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_check}
\alias{ee_check}
\title{Check if Python and the Google Earth Engine API are available on this system}
\usage{
ee_check(initialize = TRUE)
}
\arguments{
\item{initialize}{reticulate::py_available params. TRUE to attempt to initialize Python
bindings if they aren't yet available (defaults to FALSE).}
}
\description{
Check if Python and the Google Earth Engine API are available on this system
}
\examples{
library(rgee)
ee_check()
}
