% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-upload.R
\name{ee_create_asset_path}
\alias{ee_create_asset_path}
\title{Create a folder or ImageCollection into GEE assets (DEPRECATED)}
\usage{
ee_create_asset_path(path_asset, asset_type = "folder", quiet = FALSE)
}
\arguments{
\item{path_asset}{a character vector containing a single path name.}

\item{asset_type}{a character vector containing the asset type. 'folder' or 'imagecollection'.}

\item{quiet}{logical; suppress info message.}
}
\description{
Create a folder or ImageCollection into GEE assets (DEPRECATED)
}
