% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_upload_toasset.R
\name{ee_create_json}
\alias{ee_create_json}
\title{Compile the manifest (*.JSON) using import json}
\usage{
ee_create_json(manifest, path)
}
\description{
Compile the manifest (*.JSON) using import json
}
