% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_upload_toasset.R
\name{ee_create_manifest}
\alias{ee_create_manifest}
\title{Create the *.JSON (manifest) using import json}
\usage{
ee_create_manifest(x, asset_id, uris, start_time = NULL,
  end_time = NULL, properties = NULL, pyramiding_policy = "MEAN")
}
\description{
Create the *.JSON (manifest) using import json
}
