% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_create_pyenv}
\alias{ee_create_pyenv}
\title{Create an isolated Python virtual environment to be used in rgee}
\usage{
ee_create_pyenv(python_env)
}
\arguments{
\item{python_env}{The name of, or path to, a Python virtual environment.}
}
\description{
Create an isolated Python virtual environment to be used in rgee
}
\examples{
\dontrun{
library(rgee)
ee_create_pyenv('ee')
ee_discover_pyenvs()
ee_set_pyenv(
  python_path = ".../ee/bin/python",
  python_env = 'ee')
ee_install_python_packages()
}
}
