% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_discover_pyenvs}
\alias{ee_discover_pyenvs}
\title{Discover all the Python environments available
in the system}
\usage{
ee_discover_pyenvs()
}
\description{
This function enables callers to check which versions
of Python will be discovered on a system. This function
was adapted from \code{\link[reticulate]{py_discover_config}}.
}
\examples{
\dontrun{
library(rgee)
ee_create_pyenv('ee')
ee_discover_pyenvs()
ee_set_pyenv(
  python_path = ".../ee/bin/python",
  python_env = 'ee')
ee_install_python_packages()
}
}
