% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_discover_python_versions}
\alias{ee_discover_python_versions}
\title{Discover the Python versions available in the system}
\usage{
ee_discover_python_versions()
}
\description{
This function enables callers to check which versions
of Python will be discovered on a system. This function
was adapted from \code{\link[reticulate]{py_discover_config}}.
}
\examples{
\dontrun{
library(rgee)
ee_discover_python_versions()
}
}
