% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_aux.R
\name{ee_Download}
\alias{ee_Download}
\title{Download your last EE export task "COMPLETED"}
\usage{
ee_Download(task, filename = NULL, overwrite = FALSE)
}
\arguments{
\item{task}{List generated after finished correctly a EE task. See
\code{ee_Export()} for details.}

\item{filename}{Output filename.}

\item{overwrite}{A boolean indicating whether "filename" should be overwritten.}
}
\description{
Download your last EE export task "COMPLETED"
}
