% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_download.R
\name{ee_download_drive}
\alias{ee_download_drive}
\title{Move the results of an Earth Engine task to export from Google Drive to Hard disk}
\usage{
ee_download_drive(task, filename = NULL, overwrite = FALSE)
}
\arguments{
\item{task}{List generated after finished correctly a EE task. See
\code{ee_Export()} for details.}

\item{filename}{Output filename.}

\item{overwrite}{A boolean indicating whether "filename" should be overwritten.}
}
\description{
Download an EE export task saved in Google Drive.
}
\examples{
\dontrun{
 library(rgee)
 library(stars)
 library(sf)
 library(mapview)

 ee_initialize()

 # Communal Reserve Amarakaeri - Peru
 xmin <- -71.132591318
 xmax <- -70.953664315
 ymin <- -12.892451233
 ymax <- -12.731116372
 x_mean <- (xmin + xmax) / 2
 y_mean <- (ymin + ymax) / 2

 ROI <- c(xmin, ymin, xmin, ymax, xmax, ymax, xmax, ymin, xmin, ymin)
 ROI_polygon <- matrix(ROI, ncol = 2, byrow = TRUE) \%>\%
   list() \%>\%
   st_polygon() \%>\%
   st_sfc() \%>\%
   st_set_crs(4326)
 ee_geom <- ee_as_eegeom(ROI_polygon)

 # Get the mean annual NDVI for 2011
 cloudMaskL457 <- function(image) {
   qa <- image$select("pixel_qa")
   cloud <- qa$bitwiseAnd(32L)$
     And(qa$bitwiseAnd(128L))$
     Or(qa$bitwiseAnd(8L))
   mask2 <- image$mask()$reduce(ee_Reducer()$min())
   image <- image$updateMask(cloud$Not())$updateMask(mask2)
   image$normalizedDifference(list("B4", "B3"))
 }

 ic_l5 <- ee_ImageCollection("LANDSAT/LT05/C01/T1_SR")$
   filterBounds(ee_geom)$
   filterDate("2011-01-01", "2011-12-31")$
   map(cloudMaskL457)
 mean_l5 <- ic_l5$mean()$rename("NDVI")
 mean_l5 <- mean_l5$reproject(crs = "EPSG:4326", scale = 500)
 mean_l5_Amarakaeri <- mean_l5$clip(ee_geom)

 # Download -> ee$Image
 task_img <- ee$batch$Export$image$toDrive(
   image = mean_l5_Amarakaeri,
   description = "Amarakaeri_task_1",
   folder = "Amarakaeri",
   fileFormat = "GEOTIFF",
   fileNamePrefix = "NDVI_l5_2011_Amarakaeri"
 )
 task_img$start()
 ee_download_monitoring(task_img, eeTaskList = TRUE) # optional
 img <- ee_download_drive(
   task = task_img,
   filename = "amarakaeri.tif",
   overwrite = TRUE
 )
 plot(img)

 # Download -> ee$FeatureCollection
 amk_f <- ee$FeatureCollection(list(ee$Feature(ee_geom, list(name = "Amarakaeri"))))
 amk_fc <- ee$FeatureCollection(amk_f)
 task_vector <- ee$batch$Export$table$toDrive(
   collection = amk_fc,
   description = "Amarakaeri_task_2",
   folder = "Amarakaeri",
   fileFormat = "GEOJSON",
   fileNamePrefix = "geometry_Amarakaeri"
 )

 task_vector$start()
 ee_download_monitoring(task_vector, eeTaskList = TRUE) # optional
 amk_geom <- ee_download_drive(
   task = task_vector,
   filename = "amarakaeri.geojson",
   overwrite = TRUE
 )

 plot(amk_geom$geometry, add = TRUE, border = "red", lwd = 3)
}
}
