% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_download.R
\name{ee_download_gcs}
\alias{ee_download_gcs}
\title{Move the results of an Earth Engine task to export from Google Cloud Storage to Hard disk}
\usage{
ee_download_gcs()
}
\description{
Download an EE export task saved in Google Cloud Storage.
}
