% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_download.R
\name{ee_download_monitoring}
\alias{ee_download_monitoring}
\title{Monitoring Earth Engine task progress}
\usage{
ee_download_monitoring(task, eeTaskList = FALSE)
}
\arguments{
\item{task}{List generated after finished correctly a EE task. See
\code{ee_Export()} for details.}

\item{eeTaskList}{Logical, if \code{TRUE}, all Earth Engine tasks will listing initially.}
}
\description{
Monitoring Earth Engine task progress
}
\examples{
\dontrun{
library(rgee)
ee_initialize()

ee_download_monitoring(eelist=TRUE)
}
}
