% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_earthengine_upgrade}
\alias{ee_earthengine_upgrade}
\title{Upgrade the Earth Engine Python API}
\usage{
ee_earthengine_upgrade(
  method = c("auto", "virtualenv", "conda"),
  conda = "auto",
  envname = NULL,
  pip = TRUE,
  conda_python_version = "3.7",
  quiet = FALSE,
  ...
)
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically
finds a method that will work in the local environment. Change the
default to force a specific installation method. Note that the
"virtualenv" method is not available on Windows (as this isn't
supported by rgee). Note also that since this command runs
without privilege the "system" method is available only on Windows.}

\item{conda}{Path to conda executable (or "auto" to find conda
using the PATH and other conventional install locations).}

\item{envname}{Name of Python environment, or full path, which Python
packages are to be installed.}

\item{pip}{Logical. Use pip for package installation? This
is only relevant when Conda environments are used, as
otherwise packages will be installed from the
Conda repositories.}

\item{conda_python_version}{the Python version installed in the
created conda environment. Python 3.7 is installed by default.}

\item{quiet}{logical. Suppress info message}

\item{...}{other arguments passed to
\code{reticulate::\link[reticulate:conda_install]{conda_install}}
or
\code{reticulate::\link[reticulate:virtualenv_install]{virtualenv_install}}.}
}
\description{
This function upgrade the Earth Engine Python API (earthengine-api)
to the latest version.
}
\details{
It is neccessary restart R to observe change when
installing Python packages. rgee only is compatible with Python
version 3.5 >=.
}
\examples{
\dontrun{
library(rgee)
ee_earthengine_upgrade()
}
}
