% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_functions.R
\name{ee_FeatureCollection}
\alias{ee_FeatureCollection}
\title{Soy una bonita funcion que no hace nada 3 :)}
\usage{
ee_FeatureCollection(ID)
}
\arguments{
\item{ID}{holi2}
}
\description{
Soy una bonita funcion que no hace nada 3 :)
}
