% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{ee_Filter}
\alias{ee_Filter}
\title{An object to represent collection filters.}
\usage{
ee_Filter(opt_filter = NULL)
}
\arguments{
\item{opt_filter}{Optional filter to add.}
}
\value{
Filter EarthEngine Python class
}
\description{
An object to represent collection filters.
}
\details{
This constructor accepts the following args:
\enumerate{
\item Another filter.
\item A list of filters (which are implicitly AND ed together).
\item A ComputedObject returning a filter. Users should not be making these;
they are produced by the generator functions below.
}
}
\section{Methods}{

\itemize{
\item \strong{always():}
\item \strong{And():}
\item \strong{aside():}
\item \strong{calendarRange():}
\item \strong{contains():}
\item \strong{date():}
\item \strong{dateRangeContains():}
\item \strong{dayOfYear():}
\item \strong{disjoint():}
\item \strong{encode():}
\item \strong{eq():}
\item \strong{equals():}
\item \strong{freeze():}
\item \strong{geometry():}
\item \strong{getInfo():}
\item \strong{greaterThan():}
\item \strong{greaterThanOrEquals():}
\item \strong{gt():}
\item \strong{gte():}
\item \strong{hasGeometry():}
\item \strong{initialize():}
\item \strong{inList():}
\item \strong{intersects():}
\item \strong{isContained():}
\item \strong{isVariable():}
\item \strong{lessThan():}
\item \strong{lessThanOrEquals():}
\item \strong{listContains():}
\item \strong{lt():}
\item \strong{lte():}
\item \strong{maxDifference():}
\item \strong{metadata_():}
\item \strong{name():}
\item \strong{neq():}
\item \strong{never():}
\item \strong{Not():}
\item \strong{notEquals():}
\item \strong{notNull():}
\item \strong{Or():}
\item \strong{predicateCount():}
\item \strong{rangeContains():}
\item \strong{reset():}
\item \strong{serialize():}
\item \strong{stringContains():}
\item \strong{stringEndsWith():}
\item \strong{stringStartsWith():}
\item \strong{withinDistance():}
}
}

