% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-upload.R
\name{ee_gcs_to_asset_table}
\alias{ee_gcs_to_asset_table}
\title{Move a zipped shapefile from GCS to EE asset}
\usage{
ee_gcs_to_asset_table(gs_uri, asset_id, quiet = FALSE)
}
\arguments{
\item{gs_uri}{Character. It represents the full name of an
zipped shapefile in a GCS bucket.}

\item{asset_id}{Character. What to call the file once uploaded
to the Earth Engine Asset}

\item{quiet}{Logical. Suppress info message.}
}
\description{
Pass a zipped shapefile of gcs to Earth Engine Asset
}
\examples{
\dontrun{
library(rgee)
library(sf)
ee_Initialize(gcs = TRUE)

# Create sf object
nc <- st_read(system.file("shape/nc.shp", package="sf"))
asset_id <- sprintf("\%s/\%s",ee_get_assethome(),'sf_nc')

# Method 1
# 1. Pass the sf to a zip file
zipfile <- create_shp_zip(nc)

# 2. From local to gcs
gs_uri <- ee_local_to_gcs(x = zipfile, bucket = 'rgee_dev')

# 3. Pass the sf to a zip file
ee_gcs_to_asset_table(
  gs_uri = gs_uri,
  asset_id = asset_id
)

# OPTIONAL: Monitoring progress
ee_monitoring()

# OPTIONAL: Display results
ee_sf_01 <- ee$FeatureCollection(asset_id)
Map$centerObject(ee_sf_01)
Map$addLayer(ee_sf_01)

# Method 2
ee_sf_02 <- sf_as_ee(x = nc,
                     assetId = asset_id,
                     bucket = "rgee_dev",
                     via = 'gcs')
Map$centerObject(ee_sf_02)
Map$addLayer(ee_sf_02)
}
}
