% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-upload.R
\name{ee_gcs_to_image}
\alias{ee_gcs_to_image}
\title{Move a GeoTIFF image from GCS to EE asset}
\usage{
ee_gcs_to_image(
  x,
  gs_uri,
  asset_id,
  properties = NULL,
  start_time = "1970-01-01",
  end_time = "1970-01-01",
  pyramiding_policy = "MEAN",
  quiet = FALSE
)
}
\arguments{
\item{x}{stars object.}

\item{gs_uri}{Character. It represents the full name of the
GeoTIFF file in a GCS bucket.}

\item{asset_id}{Character. What to call the file once uploaded
to the Earth Engine Asset. e.g. users/datacolecfbf/mydatacollection.}

\item{properties}{List. Set of parameters to be set up as properties
of the EE object.}

\item{start_time}{Character. Sets the start time property to a number
or date.}

\item{end_time}{Character. Sets the end time property to a number
or date.}

\item{pyramiding_policy}{The pyramid reduction policy to use.}

\item{quiet}{Logical. Suppress info message.}
}
\description{
Pass a GeoTIFF image of gcs to Earth Engine Asset
}
\examples{
\dontrun{
library(rgee)
library(stars)
ee_Initialize(gcs = TRUE)

# Get the filename of a image
tif <- system.file("tif/L7_ETMs.tif", package = "stars")
x <- read_stars(tif)
asset_id <- sprintf("\%s/\%s",ee_get_assethome(),'stars_l7')

# Method 1
# 1. Move from local to gcs
gs_uri <- ee_local_to_gcs(x = tif, bucket = 'rgee_dev')

# 2. Pass from gcs to asset
ee_gcs_to_image(
  x = x,
  gs_uri = gs_uri,
  asset_id = asset_id
)

# OPTIONAL: Monitoring progress
ee_monitoring()

# OPTIONAL: Display results
ee_stars_01 <- ee$Image(asset_id)
Map$centerObject(ee_stars_01)
Map$addLayer(ee_stars_01)

# Method 2
ee_sf_02 <- stars_as_ee(x = x,
                        assetId = asset_id,
                        bucket = "rgee_dev")
Map$centerObject(ee_sf_02)
Map$addLayer(ee_sf_02)
}
}
