% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{ee_Geometry}
\alias{ee_Geometry}
\title{Constructs an Earth Engine geometry.}
\usage{
ee_Geometry(geo_json = NULL, opt_proj = NULL, opt_geodesic = NULL,
  opt_evenOdd = NULL)
}
\arguments{
\item{geo_json}{The GeoJSON object describing the geometry or a
computed object to be reinterpred as a Geometry. Supports
CRS specifications as per the GeoJSON spec, but only allows named
(rather than "linked" CRSs). If this includes a 'geodesic' field,
and opt_geodesic is not specified, it will be used as opt_geodesic.}

\item{opt_proj}{An optional projection specification, either as an
ee.Projection, as a CRS ID code or as a WKT string. If specified,
overrides any CRS found in the geo_json parameter. If unspecified and
the geo_json does not declare a CRS, defaults to "EPSG:4326"
(x=longitude, y=latitude).}

\item{opt_geodesic}{Whether line segments should be interpreted as spherical
geodesics. If false, indicates that line segments should be
interpreted as planar lines in the specified CRS. If absent,
defaults to true if the CRS is geographic (including the default
EPSG:4326), or to false if the CRS is projected.}

\item{opt_evenOdd}{If true, polygon interiors will be determined by the even/odd
rule, where a point is inside if it crosses an odd number of edges to
reach a point at infinity. Otherwise polygons use the left-inside
rule, where interiors are on the left side of the shell's edges when
walking the vertices in the given order. If unspecified, defaults to
True.}
}
\description{
Constructs an Earth Engine geometry.
}
