% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_get.R
\name{ee_get_assethome}
\alias{ee_get_assethome}
\title{Get the Asset home name}
\usage{
ee_get_assethome()
}
\value{
Character. The name of the Earth Engine Asset home
(e.g. users/datacolecfbf)
}
\description{
Get the Asset home name
}
\examples{
\dontrun{
library(rgee)
ee_Initialize()
ee_get_assethome()
}
}
\seealso{
Other path utils: 
\code{\link{ee_get_earthengine_path}()}
}
\concept{path utils}
