% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Date.R
\name{ee_get_date}
\alias{ee_get_date}
\title{Get the date of a Earth Engine Image}
\usage{
ee_get_date(image, time_end = FALSE)
}
\arguments{
\item{image}{The Earth Engine Image}

\item{time_end}{Logical. If TRUE, the
system:time_end property will also be returned.
See details.}
}
\description{
Get the date of a Earth Engine Image
}
\details{
The properties system:time_start and system:time_end represent the
image dates based on a timestamp in milliseconds
since the UNIX epoch. The system:time_start set the
start period while system:time_end the end period. See
this \href{https://developers.google.com/earth-engine/glossary}{link}
for more information.
}
\examples{
library(rgee)
ee_Initialize()
ee_reattach()
l8 <- ee$Image('LANDSAT/LC08/C01/T1_TOA/LC08_044034_20140318')
ee_get_date(l8)
srtm <- ee$Image('CGIAR/SRTM90_V4')
ee_get_date(srtm, time_end = TRUE)
}
