% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_get.R
\name{ee_get_earthengine_path}
\alias{ee_get_earthengine_path}
\title{Get the path where the credentials are stored}
\usage{
ee_get_earthengine_path()
}
\value{
A character that represents the path credential of a specific
user
}
\description{
Get the path where the credentials are stored
}
\seealso{
Other path utils: 
\code{\link{ee_get_assethome}()}
}
\concept{path utils}
