% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Date.R
\name{ee_get_img_date}
\alias{ee_get_img_date}
\title{Get the date of a EE Image}
\usage{
ee_get_img_date(x, time_end = FALSE)
}
\arguments{
\item{x}{EE Image or ImageCollection object}

\item{time_end}{Logical. If TRUE, the system:time_end property will
also be returned. See details.}
}
\value{
An List object with the id, time_start and time_end
(if \code{time_end} argument is TRUE) of the image.
}
\description{
Get the date of a EE Image
}
\details{
\code{system:time_start} set the start period of data acquisition while
\code{system:time_end} does the same for the end period. See this
\href{https://developers.google.com/earth-engine/glossary}{link} for
getting more information.
}
\examples{
\dontrun{
library(rgee)
ee_reattach()
ee_Initialize()
l8 <- ee$Image('LANDSAT/LC08/C01/T1_TOA/LC08_044034_20140318')
ee_get_img_date(l8)
srtm <- ee$Image('CGIAR/SRTM90_V4')
ee_get_img_date(srtm, time_end = TRUE)
}
}
