% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_image.R
\name{ee_image_as_raster}
\alias{ee_image_as_raster}
\title{Convert an Earth Engine (EE) image into a raster object}
\usage{
ee_image_as_raster(
  image,
  region,
  dsn = NULL,
  via = "getInfo",
  scale = NULL,
  maxPixels = 1e+09,
  container = "rgee_backup",
  quiet = FALSE
)
}
\arguments{
\item{image}{ee$Image to be converted into a raster object}

\item{region}{EE Geometry Rectangle (ee$Geometry$Rectangle). The
CRS needs to be the same that the x argument otherwise it will be
forced. If not specified, image bounds will be taken.}

\item{dsn}{Character. Output filename. If missing,
\code{ee_image_as_character} will create a temporary file.}

\item{via}{Character. Method to fetch data about the object. Multiple
options supported. See details.}

\item{scale}{Numeric. The resolution in meters per pixel. Defaults
to the native resolution of the image assset.}

\item{maxPixels}{Numeric. The maximum allowed number of pixels in the
exported image. The task will fail if the exported region covers
more pixels in the specified projection. Defaults to 100,000,000.}

\item{container}{Character. Name of the folder ('drive') or bucket ('gcs')
to be exported into (ignored if \code{via} is not defined as "drive" or
"gcs").}

\item{quiet}{logical. Suppress info message}
}
\value{
A character object
}
\description{
Convert an Earth Engine (EE) image into a raster object
}
\details{
\code{ee_image_as_raster} supports the download of \code{ee$Image}
by three different options: "getInfo", "drive", and "gcs". When "getInfo"
is set in the \code{via} argument, \code{ee_image_as_raster} will make an
REST call to retrieve all the known information about the object. The
advantage of use "getInfo" is a direct and faster download. However, there
is a limitation of 5000 features by request which makes it not recommendable
for large collections. Instead of "getInfo", the options: "drive" and "gcs"
are suitable for large collections since they use an intermediate container.
They are Google Drive and Google Cloud Storage respectively. For getting
more information about exporting data take a look at the
\href{https://developers.google.com/earth-engine/exporting}{Google Earth
Engine Guide - Export data}.
}
