% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_download.R
\name{ee_image_to_drive}
\alias{ee_image_to_drive}
\title{Creates a task to export an EE Image to Drive.}
\usage{
ee_image_to_drive(image, description = "myExportImageTask",
  folder = NULL, fileNamePrefix = NULL, dimensions = NULL,
  region = NULL, scale = NULL, crs = NULL, crsTransform = NULL,
  maxPixels = NULL, shardSize = NULL, fileDimensions = NULL,
  skipEmptyTiles = NULL, fileFormat = NULL, formatOptions = NULL)
}
\arguments{
\item{image}{The image to be exported.}

\item{description}{Human-readable name of the task.}

\item{folder}{The name of a unique folder in your Drive account to export into. Defaults to the root of the drive.}

\item{fileNamePrefix}{The Google Drive filename for the export. Defaults to the name of the task.}

\item{dimensions}{The dimensions of the exported image. Takes either a single positive integer as the maximum dimension or "WIDTHxHEIGHT" where WIDTH and HEIGHT are each positive integers.}

\item{region}{The lon,lat coordinates for a LinearRing or Polygon specifying the region to export. Can be specified as a nested lists of numbers or a serialized string. Defaults to the image's region.}

\item{scale}{The resolution in meters per pixel. Defaults to the native resolution of the image assset unless a crsTransform is specified.}

\item{crs}{The coordinate reference system of the exported image's projection. Defaults to the image's default projection.}

\item{crsTransform}{A comma-separated string of 6 numbers describing the affine transform of the coordinate reference system of the exported image's projection, in the order: xScale, xShearing, xTranslation, yShearing, yScale and yTranslation. Defaults to the image's native CRS transform.}

\item{maxPixels}{The maximum allowed number of pixels in the exported image. The task will fail if the exported region covers more pixels in the specified projection. Defaults to 100,000,000.}

\item{shardSize}{Size in pixels of the shards in which this image will be computed. Defaults to 256.}

\item{fileDimensions}{The dimensions in pixels of each image file, if the image is too large to fit in a single file. May specify a single number to indicate a square shape, or a list of two dimensions to indicate (width,height). Note that the image will still be clipped to the overall image dimensions. Must be a multiple of shardSize.}

\item{skipEmptyTiles}{If true, skip writing empty (i.e. fully-masked) image tiles. Defaults to false.}

\item{fileFormat}{The string file format to which the image is exported. Currently only 'GeoTIFF' and 'TFRecord' are supported, defaults to 'GeoTIFF'.}

\item{formatOptions}{A dictionary of string keys to format specific options. **kwargs: Holds other keyword arguments that may have been deprecated such as 'crs_transform', 'driveFolder', and 'driveFileNamePrefix'.}
}
\value{
An unstarted Task that exports the image to Drive.
}
\description{
Creates a task to export an EE Image to Drive.
}
