% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_functions.R
\name{ee_ImageCollection}
\alias{ee_ImageCollection}
\title{Soy una bonita funcion que no hace nada 2 :)}
\usage{
ee_ImageCollection(ID)
}
\arguments{
\item{ID}{holi}
}
\description{
Soy una bonita funcion que no hace nada 2 :)
}
