% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_install-tools}
\alias{ee_install-tools}
\alias{ee_install_drivers}
\alias{ee_install_rgee_python_packages}
\alias{ee_install_python_package}
\alias{ee_install_python_ee}
\alias{ee_remove_drivers}
\alias{ee_remove_credentials}
\title{Interface to install non-R rgee dependencies}
\usage{
ee_install_drivers(GoogleChromeVersion)

ee_install_rgee_python_packages(conda = FALSE, quiet = quiet)

ee_install_python_package(pypackage, conda = FALSE, quiet = FALSE)

ee_install_python_ee(conda = FALSE, quiet = FALSE)

ee_remove_drivers(quiet = TRUE)

ee_remove_credentials(user, quiet = TRUE)
}
\arguments{
\item{GoogleChromeVersion}{Google Chrome version of this system.}

\item{conda}{If TRUE conda is used instead of pip. conda executable needs setting in PATH.}

\item{quiet}{logical. Suppress info message}

\item{pypackage}{character. Package names to install.}

\item{user}{character. User to remove credentials}
}
\description{
R functions for installing and removing Python Selenium Chromedriver and Third-Party Python packages.
}
\details{
It is neccessary restart R to observe change when installing a Python packages. rgee only is
compatible with Python version 3.5 >.
}
\examples{
\dontrun{
library(rgee)

# Recommended way to use rgee for Linux users
virtualenv_remove("rgee")
virtualenv_create("rgee", python = "python3.7")
use_virtualenv("rgee") # Restart R

ee_check()
ee_install_rgee_python_packages() # Install rgee python packages
ee_install_drivers() # Install selenim drivers
ee_Initialize(drive=TRUE,gcs=TRUE) # Install credentials
ee_check()

# Recommended way to install external dependencies of rgee for Windows users
}
}
