% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_install}
\alias{ee_install}
\title{Create an isolated Python virtual environment with all rgee dependencies.}
\usage{
ee_install(py_env = "rgee", confirm = interactive())
}
\arguments{
\item{py_env}{Character. The name, or full path, of the Python environment
to be used by rgee.}

\item{confirm}{Logical. Confirm before restarting R?.}
}
\description{
Create an isolated Python virtual environment with all rgee dependencies.
}
\examples{
\dontrun{
library(rgee)
ee_install() #It is just necessary once
ee_Initialize()
}
}
