% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_install_ChromeDriver}
\alias{ee_install_ChromeDriver}
\title{Install ChromeDriver in the system}
\usage{
ee_install_ChromeDriver(ChromeDriverVersion)
}
\arguments{
\item{ChromeDriverVersion}{Google Chrome version of the system.}
}
\description{
Install ChromeDriver in the system
}
\examples{
\dontrun{
library(rgee)
ee_install_ChromeDriver()
}
}
\seealso{
\code{\link[rgee]{ee_upload}}
}
