% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_install_create_pyenv}
\alias{ee_install_create_pyenv}
\title{Create an isolated Python virtual environment to be used in rgee}
\usage{
ee_install_create_pyenv(python_env = "rgee")
}
\arguments{
\item{python_env}{The name of, or path to, a Python virtual environment.}
}
\value{
Character. The dirpath of the virtual environment created.
}
\description{
Create an isolated Python virtual environment to be used in rgee
}
\examples{
\dontrun{
#' library(rgee)

### rgee installation

# 1. Initialize rgee with ee_Initialize(). If there is no any Python environment, miniconda
# will be installed by default.
ee_Initialize()

# 2. Create a Python environment, e.g. ee.
pyenv <- ee_install_create_pyenv(python_env = "ee")

# Find others Python environments in the system.
# ee_install_discover_pyenvs()

# 3. Set a Python environment (e.g. ee) and restart R to see changes.
ee_install_set_pyenv(pyenv, install = TRUE)

# 4. Install Python package dependencies and restart R to see changes.
ee_install_python_packages()

# 5. Initialize rgee again!
ee_Initialize()
}
}
