% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_install_discover_pyenvs}
\alias{ee_install_discover_pyenvs}
\title{Discover all the Python environments available
in the system}
\usage{
ee_install_discover_pyenvs(use_py_discover_config = TRUE)
}
\arguments{
\item{use_py_discover_config}{Logical. If TRUE
will use \link[=reticulate]{py_discover_config} to find
versions of Python in the system.  Otherwise, will use
\link[=reticulate]{conda_list} for Window OS and
\link[=reticulate]{virtualenv_list} for Unix system.}
}
\value{
Python configuration object (reticulate).
}
\description{
This function enables callers to check which versions
of Python will be discovered on a system.
}
\examples{
\dontrun{
library(rgee)

## It is necessary just once, not mandatory

# 1. Create a Python environment, e.g. ee.
pyenv <- ee_install_create_pyenv(py_env = "ee")

# OPTIONAL: Find others Python path in the system.
# ee_install_discover_pyenvs()

# 2. Set a Python path in .Renviron (EARTHENGINE_PYTHON)
# to be used in future sessions
ee_install_set_pyenv(pyenv)

# 3. Now run ee_Initialize()
ee_Initialize()
}
}
