% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_install_earthengine_upgrade}
\alias{ee_install_earthengine_upgrade}
\title{Upgrade the Earth Engine Python API (DEPRECATED)}
\usage{
ee_install_earthengine_upgrade(
  method = c("auto", "virtualenv", "conda"),
  conda = "auto",
  pip = FALSE,
  python_version = NULL,
  confirm = interactive(),
  ...
)
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically
finds a method that will work in the local environment. Change the
default to force a specific installation method. Note that the
"virtualenv" method is not available on Windows.}

\item{conda}{The path to a conda executable. Use "auto" to allow
reticulate to automatically find an appropriate conda binary. See
\strong{Finding Conda} for more details.}

\item{pip}{Logical. Use pip for package installation? This
is only relevant when Conda environments are used, as
otherwise packages will be installed from the
Conda repositories.}

\item{python_version}{The requested Python version. Ignored when
attempting to install with a Python virtual environment.}

\item{confirm}{Logical. Confirm if restart R when the 'install'
argument is TRUE.}

\item{...}{Additional arguments passed to \link[=reticulate]{conda_install}
or \link[=reticulate]{virtualenv_install}.}
}
\description{
Upgrade the Earth Engine Python API (earthengine-api) to the latest
version. This function is a wrapper around  \code{reticulate::py_install()}.
Due to recent changes in reticulate it is no longer necessary.
ee_install_python_packages and  ee_install_earthengine_upgrade will be
removed in rgee 0.5.3
}
\details{
It is neccessary restart R to observe change when
installing Python packages. rgee only is compatible with Python
version 3.5 >=. On Linux and OS X the "virtualenv" method will
be used by default ("conda" will be used if virtualenv isn't available). On
Windows, the "conda" method is always used.
}
\examples{
\dontrun{
library(rgee)
ee_install_earthengine_upgrade()
}
}
\seealso{
\code{reticulate::py_install()}
}
