% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_install_set_pyenv}
\alias{ee_install_set_pyenv}
\title{Configure which version of Python to use with rgee}
\usage{
ee_install_set_pyenv(
  py_path,
  py_env = NULL,
  Renviron = "global",
  confirm = interactive(),
  quiet = FALSE
)
}
\arguments{
\item{py_path}{The path to a Python interpreter}

\item{py_env}{The name of the conda or venv environment. If
NULL, \code{\link{ee_install_upgrade}} and \code{\link{py_install}} functions will not work.}

\item{Renviron}{Character. If it is "global" the environment variables are set in
the .Renviron located in the Sys.getenv("HOME") folder. On the other hand,  if
it is "local" the environment variables are set in the .Renviron on the
working directory (getwd()). Finally, users can also enter a specific path
(see examples).}

\item{confirm}{Logical. Confirm before restarting R?.}

\item{quiet}{Logical. Suppress info message}
}
\value{
no return value, called for setting EARTHENGINE_PYTHON in .Renviron
}
\description{
Configure which version of Python to use with rgee. This function creates two
environment variables: 'EARTHENGINE_PYTHON' and 'EARTHENGINE_ENV' both will be
saved into the file .Renviron.
}
\examples{
\dontrun{
library(rgee)

## For macOS and Linux users
# ee_install_set_pyenv(py_path = "/usr/bin/python3", Renviron = "local")
# ee_install_set_pyenv(py_path = "/usr/bin/python3", Renviron = "/home/zgis/")

## For Window users
# ee_install_set_pyenv(py_path = "C:\\Users\\Ellen\\Anaconda3\\envs\\rgee", py_env = "rgee")
}
}
\seealso{
Other ee_install functions: 
\code{\link{ee_install_upgrade}()},
\code{\link{ee_install}()}
}
\concept{ee_install functions}
