% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_install_upgrade}
\alias{ee_install_upgrade}
\title{Upgrade the Earth Engine Python API}
\usage{
ee_install_upgrade(version = NULL)
}
\arguments{
\item{version}{Character. The Earth Engine Python API version to upgrade.
By default \code{rgee::ee_version()}.}
}
\description{
Upgrade the Earth Engine Python API
}
