% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-subsetting.R
\name{ee_length_ic}
\alias{ee_length_ic}
\alias{length.ee.imagecollection.ImageCollection}
\title{Length of an Earth Engine ImageCollection Object}
\usage{
\method{length}{ee.imagecollection.ImageCollection}(x)
}
\arguments{
\item{x}{an EE ImageCollection Object.}
}
\description{
Set the length of an Earth Engine Image.
}
