% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Map.R
\docType{data}
\name{ee_Map}
\alias{ee_Map}
\title{Function to display map in earthengine}
\format{An object of class \code{environment} of length 5.}
\usage{
ee_Map
}
\description{
Function to display map in earthengine
}
\keyword{datasets}
