% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-subsetting.R
\name{ee_name}
\alias{ee_name}
\alias{names<-.ee.image.Image}
\alias{names.ee.image.Image}
\title{Names of Earth Engine Images Layers}
\usage{
\method{names}{ee.image.Image}(x) <- value

\method{names}{ee.image.Image}(x)
}
\arguments{
\item{x}{an EE Image object.}

\item{value}{a character vector with the same length as x.}
}
\description{
Get or set the names of the layers of an Earth Engine Image object.
}
