% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_quota.R
\name{ee_quota}
\alias{ee_quota}
\title{Returns quota usage details for the asset root with the given ID.}
\usage{
ee_quota()
}
\description{
Returns quota usage details for the asset root with the given ID.
}
\examples{
\dontrun{
ee_quota()
}
}
\author{
\href{Samapriya Roy}{https://github.com/samapriya}
}
