% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_image.R
\name{ee_read_raster}
\alias{ee_read_raster}
\title{helper function to read raster (ee_as_raster)}
\usage{
ee_read_raster(img_dsn, band_names, metadata)
}
\description{
helper function to read raster (ee_as_raster)
}
