% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_image.R
\name{ee_read_stars}
\alias{ee_read_stars}
\title{helper function to read raster (ee_read_stars)}
\usage{
ee_read_stars(img_dsn, band_names, metadata)
}
\description{
helper function to read raster (ee_read_stars)
}
