% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ee_reattach}
\alias{ee_reattach}
\title{Reattach ee as a reserved word}
\usage{
ee_reattach()
}
\description{
Reattach ee as a reserved word
}
\seealso{
Other session management functions: 
\code{\link{ee_Initialize}()},
\code{\link{ee_user_info}()},
\code{\link{ee_users}()},
\code{\link{ee_version}()}
}
\concept{session management functions}
