% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_remove.R
\name{ee_remove_ChromeDriver}
\alias{ee_remove_ChromeDriver}
\title{Delete chromedriver}
\usage{
ee_remove_ChromeDriver(quiet = FALSE)
}
\arguments{
\item{quiet}{Logical (optional). Suppress info messages.}
}
\description{
Delete chromedriver
}
\examples{
\dontrun{
library(rgee)
ee_remove_ChromeDriver()
}
}
