% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_remove.R
\name{ee_remove_driver}
\alias{ee_remove_driver}
\title{Delete chromedriver}
\usage{
ee_remove_driver(quiet = FALSE)
}
\arguments{
\item{quiet}{Logical (optional). Suppress info messages.}
}
\description{
Delete chromedriver
}
\examples{
{
ee_remove_driver()
}
}
