% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_manage.R
\name{ee_remove_project_chr}
\alias{ee_remove_project_chr}
\title{Remove EE projects info}
\usage{
ee_remove_project_chr(x)
}
\arguments{
\item{x}{Character (path_asset)}
}
\description{
Remove EE projects info
}
