% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_search.R
\name{ee_search-tools}
\alias{ee_search-tools}
\alias{ee_dataset}
\alias{ee_search_startdate}
\alias{ee_search_enddate}
\alias{ee_search_type}
\alias{ee_search_provider}
\alias{ee_search_provider_list}
\alias{ee_search_tags}
\alias{ee_search_title}
\alias{ee_search_tagstitle}
\alias{ee_search_title_list}
\alias{ee_search_display}
\title{Interface to search into the Earth Engine Data Catalog}
\usage{
ee_dataset(quiet = FALSE, upgrade = FALSE)

ee_search_startdate(ee_dataset, stardate)

ee_search_enddate(ee_dataset, enddate = Sys.Date())

ee_search_type(ee_dataset, type)

ee_search_provider(ee_dataset, provider)

ee_search_provider_list(ee_dataset)

ee_search_tags(ee_dataset, ..., logical_operator = "OR")

ee_search_title(ee_dataset, ..., logical_operator = "OR")

ee_search_tagstitle(ee_dataset, ..., logical_operator = "OR")

ee_search_title_list(ee_dataset)

ee_search_display(ee_dataset, maxdisplay = 10)
}
\arguments{
\item{quiet}{logical. Suppress info message}

\item{upgrade}{Logical. If the dataset needs to be upgraded.}

\item{ee_dataset}{data.frame. Generated by rgee::ee_search_Datasets().
It is automatically updated and expanded weekly.}

\item{stardate}{Character. Start date of dataset availability.}

\item{enddate}{Character. End date of dataset availability.}

\item{type}{Character. "Image", "ImageCollection" or a "table".}

\item{provider}{Character. Name of the dataset's provider. See
ee_search_provider_list()}

\item{...}{Character vector. tags}

\item{logical_operator}{Character. Available just for rgee::ee_search_tags
and rgee::ee_search_title. 'AND' represents inclusiveness between tags in
searching and 'OR' exclusiveness.}

\item{maxdisplay}{Numeric. Maximum number of tabs to display in their browser}
}
\value{
a data.frame.
}
\description{
R functions for searching in Earth Engine's public data archive.
}
\examples{
\dontrun{
library(rgee)

ee_reattach() # reattach ee as a reserved word
ee_Initialize()

# ee_search_provider_list()
# ee_search_title_list()
myquery <- ee_dataset() \%>\%
  ee_search_type("Image") \%>\%
  ee_search_provider("WWF") \%>\%
  ee_search_tags("srtm", "flow", "direction", "dem") \%>\%
  ee_search_title("15", "Flow", logical_operator = "AND")
print(myquery$id)
}
}
