% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_set_pyenv}
\alias{ee_set_pyenv}
\title{Set the Python environment to be used on rgee}
\usage{
ee_set_pyenv(python_path, python_env)
}
\arguments{
\item{python_path}{The path to a Python interpreter, to be used with rgee.}

\item{python_env}{The name of, or path to, a Python virtual environment.}
}
\description{
Set the Python environment to be used on rgee
}
\details{
It is necessary to restart R to observe change when setting a
different Python version. ee_set_python_version will ask you to restart R.
}
\examples{
\dontrun{
library(rgee)
ee_reattach() # reattach ee as a reserved word
ee_discover_python_versions()
ee_set_python_version(
  python_path = "user/.virtualenvs/ee/bin/python",
  python_env = 'ee'
)
}
}
