% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_set_python_version}
\alias{ee_set_python_version}
\title{Set the python version to be used on rgee}
\usage{
ee_set_python_version(python_path)
}
\arguments{
\item{python_path}{The path to a Python interpreter, to be used with rgee.}
}
\description{
Set the python version to be used on rgee
}
\details{
It is necessary to restart R to observe change when setting a
different Python version. ee_set_python_version will ask you to restart R.
}
\examples{
\dontrun{
library(rgee)
ee_discover_python_versions()
ee_set_python_version(python_path = "/usr/bin/my_python_version")
}
}
