% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_download.R
\name{ee_table_to_asset}
\alias{ee_table_to_asset}
\title{Creates a task to export a FeatureCollection to an EE table asset.}
\usage{
ee_table_to_asset(collection, description = "myExportTableTask",
  assetId = NULL)
}
\arguments{
\item{collection}{The feature collection to be exported.}

\item{description}{Human-readable name of the task.}

\item{assetId}{The destination asset ID. **kwargs: Holds other keyword arguments that may have been deprecated.}
}
\value{
An unstarted Task that exports the table.
}
\description{
Creates a task to export a FeatureCollection to an EE table asset.
}
