% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_download.R
\name{ee_table_to_drive}
\alias{ee_table_to_drive}
\title{Creates a task to export a FeatureCollection to Google Cloud Storage.}
\usage{
ee_table_to_drive(collection, description = "myExportTableTask",
  folder = NULL, fileNamePrefix = NULL, fileFormat = NULL,
  selectors = NULL)
}
\arguments{
\item{collection}{The feature collection to be exported.}

\item{description}{Human-readable name of the task.}

\item{folder}{The name of a unique folder in your Drive account to export into. Defaults to the root of the drive.}

\item{fileNamePrefix}{The Google Drive filename for the export. Defaults to the name of the task.}

\item{fileFormat}{The output format: "CSV" (default), "GeoJSON", "KML", "KMZ", "SHP", or "TFRecord".}

\item{selectors}{The list of properties to include in the output, as a list of strings or a comma-separated string. By default, all properties are included. **kwargs: Holds other keyword arguments that may have been deprecated such as 'driveFolder' and 'driveFileNamePrefix'.}
}
\value{
An unstarted Task that exports the table.
}
\description{
Creates a task to export a FeatureCollection to Google Cloud Storage.
}
