% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_download.R
\name{ee_table_to_gcs}
\alias{ee_table_to_gcs}
\title{Creates a task to export a FeatureCollection to Google Cloud Storage.}
\usage{
ee_table_to_gcs(collection, description = "myExportTableTask",
  bucket = NULL, fileNamePrefix = NULL, fileFormat = NULL,
  selectors = NULL)
}
\arguments{
\item{collection}{The feature collection to be exported.}

\item{description}{Human-readable name of the task.}

\item{bucket}{The name of a Cloud Storage bucket for the export.}

\item{fileNamePrefix}{Cloud Storage object name prefix for the export. Defaults to the name of the task.}

\item{fileFormat}{The output format: "CSV" (default), "GeoJSON", "KML", "KMZ", "SHP", or "TFRecord".}

\item{selectors}{The list of properties to include in the output, as a list of strings or a comma-separated string. By default, all properties are included. **kwargs: Holds other keyword arguments that may have been deprecated such as 'outputBucket'.}
}
\value{
An unstarted Task that exports the table.
}
\description{
Creates a task to export a FeatureCollection to Google Cloud Storage.
}
