% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_upload.R
\name{ee_upload}
\alias{ee_upload}
\alias{ee_upload.character}
\alias{ee_upload.sf}
\alias{ee_upload.stars}
\alias{ee_upload.stars_proxy}
\title{Upload sf or stars objects to GEE asset}
\usage{
ee_upload(x, ...)

\method{ee_upload}{character}(x, ..., asset_filenames, properties = NULL,
  pyramiding_policy = "MEAN", one_by_one = TRUE, bucket = NULL,
  selenium_params = getOption("rgee.selenium.params"), quiet = FALSE)

\method{ee_upload}{sf}(x, ..., gmail_account, gee_asset_file,
  cache = TRUE, quiet = FALSE)

\method{ee_upload}{stars}(x, ..., gmail_account, gee_asset_file,
  type = "Float32", properties = NULL, pyramiding_policy = "MEAN",
  one_by_one = TRUE, cache = TRUE, quiet = FALSE)

\method{ee_upload}{stars_proxy}(x, ..., gmail_account, asset_filename,
  properties = NULL, pyramiding_policy = "MEAN", one_by_one = TRUE,
  cache = TRUE, quiet = FALSE)
}
\arguments{
\item{x}{character, sf or stars object.}

\item{...}{ignored}

\item{properties}{data.frame; \href{https://developers.google.com/earth-engine/image_manifest}{Manifest upload}.
See details}

\item{pyramiding_policy}{character; determines how each pixel at a given level of the pyramid is
computed from the aggregation. "MEAN" (default), "MODE" and "SAMPLE".}

\item{one_by_one}{TODO}

\item{quiet}{logical; suppress info message}

\item{gmail_account}{character. Google account name.}

\item{cache}{TODO}

\item{type}{TODO}

\item{asset_filename}{asset filename}

\item{gee_asset_path}{character. images destination path, full path for upload to Google Earth Engine,
e.g. users/pinkiepie/myponycollection.}

\item{showpassword}{TODO}
}
\description{
Send either images or vectors to Google Earth Engine asset using Selenium
}
\details{
TALKING ABOUT SELENIUM
}
\examples{
\dontrun{
library(rgee)
}
}
