% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-upload.R
\name{ee_upload_file_to_gcs}
\alias{ee_upload_file_to_gcs}
\title{Upload local files to google cloud storage}
\usage{
ee_upload_file_to_gcs(x, bucket = NULL,
  selenium_params = getOption("rgee.selenium.params"), quiet = FALSE)
}
\arguments{
\item{x}{filename (character), sf or stars object.}

\item{bucket}{bucketname you are uploading to}

\item{selenium_params}{Optional parameters when bucket is NULL. See details}

\item{quiet}{logical; suppress info message}
}
\value{
\itemize{
\item upload_url: URL for the new :class:\code{Request} object.
\item destination_path: Is the destination path in the earth engine asset.
\item asset_filename: Filename (inside earth engine)
}
}
\description{
Upload images or tables into google cloud storage for EE asset ingestion tasks.
}
\details{
laalla
}
