% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_upload_toasset.R
\name{ee_upload_istars_to_asset}
\alias{ee_upload_istars_to_asset}
\title{Upload a Image to the asset}
\usage{
ee_upload_istars_to_asset(x, type, user, destination_path, metadata)
}
\description{
Upload a Image to the asset
}
