% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_upload_toasset.R
\name{ee_upload_toasset}
\alias{ee_upload_toasset}
\title{Upload a either an image or vector into Google Earth Engine using Selenium}
\usage{
ee_upload_toasset(x, user, destination_path, metadata, nodata)
}
\description{
Upload a either an image or vector into Google Earth Engine using Selenium
}
\author{
Cesar Aybar
}
