% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Initialize.R
\name{ee_user_info}
\alias{ee_user_info}
\title{Display the credentials and general info of the initialized user}
\usage{
ee_user_info(quiet = FALSE)
}
\arguments{
\item{quiet}{Logical. Suppress info messages.}
}
\description{
Display the credentials and general info of the initialized user
}
\examples{
\dontrun{
library(rgee)
ee_reattach() # reattach ee as a reserved word
ee_Initialize()
ee_user_info()
}
}
\seealso{
Other session management functions: 
\code{\link{ee_Initialize}()},
\code{\link{ee_reattach}()},
\code{\link{ee_users}()},
\code{\link{ee_version}()}
}
\concept{session management functions}
