% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-upload.R
\name{ee_utils_create_json}
\alias{ee_utils_create_json}
\title{Convert a R list into a JSON file}
\usage{
ee_utils_create_json(x)
}
\arguments{
\item{x}{List to convert into a JSON file.}
}
\value{
A JSON file saved in a /tmp dir.
}
\description{
Convert a R list into a JSON file
}
\examples{
library(rgee)
ee_utils_create_json(list(a=10,b=10))
}
