% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_utils.R
\name{ee_utils_dataset_display}
\alias{ee_utils_dataset_display}
\title{Search into the Earth Engine Data Catalog}
\usage{
ee_utils_dataset_display(ee_search_dataset)
}
\arguments{
\item{ee_search_dataset}{Character that represents the EE dataset ID.}
}
\value{
No return value, called for displaying the Earth Engine dataset in the browser.
}
\description{
Search into the Earth Engine Data Catalog
}
\examples{
\dontrun{
 library(rgee)

 ee_datasets <- c("WWF/HydroSHEDS/15DIR", "WWF/HydroSHEDS/03DIR")
 ee_utils_dataset_display(ee_datasets)
}
}
