% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-upload.R
\name{ee_utils_get_crs}
\alias{ee_utils_get_crs}
\title{Convert EPSG, ESRI or SR-ORG code into a OGC WKT}
\usage{
ee_utils_get_crs(code)
}
\arguments{
\item{code}{The projection code.}
}
\value{
A character which represents the same projection in WKT2 string.
}
\description{
Convert EPSG, ESRI or SR-ORG code into a OGC WKT
}
\examples{
\dontrun{
library(rgee)

ee_utils_get_crs("SR-ORG:6864")
ee_utils_get_crs("EPSG:4326")
ee_utils_get_crs("ESRI:37002")
}
}
