% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_utils.R
\name{ee_utils_sak_validate}
\alias{ee_utils_sak_validate}
\title{Validate a Service account key (SaK)}
\usage{
ee_utils_sak_validate(sakfile, quiet = FALSE)
}
\arguments{
\item{sakfile}{Character. SaK filename.}

\item{quiet}{Logical. Suppress info message}
}
\description{
Validate a Service account key (SaK). local_to_gcs, raster_as_ee,
stars_as_ee, and sf_as_ee(via = "gcs_to_asset", ...) need that the SaK
have privileges to write/read objects in a GCS bucket.
}
\examples{
\dontrun{
library(rgee)

ee_Initialize()

sakfile <- "/home/rgee_dev/sak_file.json"
ee_utils_sak_validate(sakfile_l)

}
}
