% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_utils.R
\name{ee_utils_shp_to_zip}
\alias{ee_utils_shp_to_zip}
\title{Create a zip file from a sf object}
\usage{
ee_utils_shp_to_zip(
  x,
  filename,
  SHP_EXTENSIONS = c("dbf", "prj", "shp", "shx")
)
}
\arguments{
\item{x}{sf object}

\item{filename}{data source name}

\item{SHP_EXTENSIONS}{file extension of the files to save
into the zip file. By default: "dbf", "prj", "shp", "shx".}
}
\value{
Character. The full path of the zip file created.
}
\description{
Create a zip file from a sf object
}
\examples{
\dontrun{
library(rgee)
library(sf)
ee_Initialize(gcs = TRUE)

# Create sf object
nc <- st_read(system.file("shape/nc.shp", package="sf"))
assetId <- sprintf("\%s/\%s",ee_get_assethome(),'sf_nc')

# Method 1
# 1. Pass the sf to a zip file
zipfile <- ee_utils_shp_to_zip(nc)

# 2. From local to gcs
gs_uri <- local_to_gcs(x = zipfile, bucket = 'rgee_dev')

# 3. Pass the sf to a zip file
gcs_to_ee_table(
  gs_uri = gs_uri,
  assetId = assetId
)

# OPTIONAL: Monitoring progress
ee_monitoring()

# OPTIONAL: Display results
ee_sf_01 <- ee$FeatureCollection(assetId)
Map$centerObject(ee_sf_01)
Map$addLayer(ee_sf_01)

# Method 2
ee_sf_02 <- sf_as_ee(x = nc,
                     assetId = assetId,
                     bucket = "rgee_dev")
Map$centerObject(ee_sf_02)
Map$addLayer(ee_sf_02)
}
}
\seealso{
Other ee_utils functions: 
\code{\link{ee_utils_py_to_r}()},
\code{\link{ee_utils_pyfunc}()}
}
\concept{ee_utils functions}
