% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_version.R
\name{ee_version}
\alias{ee_version}
\title{Earth Engine API version}
\usage{
ee_version()
}
\value{
Character. Earth Engine Python API version used to build rgee.
}
\description{
Earth Engine API version
}
\seealso{
Other session management functions: 
\code{\link{ee_Initialize}()},
\code{\link{ee_user_info}()},
\code{\link{ee_users}()}
}
\concept{session management functions}
