% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Date.R
\name{eeDate_to_r}
\alias{eeDate_to_r}
\title{Pass Earth Engine date to R date object}
\usage{
eeDate_to_r(eedate)
}
\arguments{
\item{eedate}{EE date object (ee$Date)}
}
\description{
Pass Earth Engine and return a R date object
}
\examples{
ee_Initialize()
eeDate <- ee$Date('1980-01-01')
eeDate_to_r(eeDate)
}
