% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-math.R
\name{Math-methods}
\alias{Math-methods}
\alias{Math.ee.image.Image}
\title{Mathematical functions}
\usage{
\method{Math}{ee.image.Image}(x, ...)
}
\arguments{
\item{x}{ee$Image}

\item{...}{Ignored

Generic mathematical functions that can be used with an \code{ee$Image}
object as argument: \code{abs}, \code{sign}, \code{sqrt}, \code{ceiling},
\code{cumprod}, \code{cumsum}, \code{log}, \code{log10}, \code{log1p},
\code{log2}, \code{acos}, \code{floor}, \code{asin}, \code{atan}, \code{exp},
\code{expm1}, \code{cos}, \code{cosh}, \code{sin}, \code{sinh},
\code{tan}, and \code{tanh}.}
}
\description{
Mathematical functions
}
