% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_operators.R
\name{null-default}
\alias{null-default}
\alias{+.EarthEngineMap}
\alias{|.EarthEngineMap}
\alias{|,}
\alias{EarthEngineMap,}
\alias{EarthEngineMap-method}
\title{EarthEngineMap + EarthEngineMap; adds data from the second map to the first}
\usage{
\method{+}{EarthEngineMap}(e1, e2)

\method{|}{EarthEngineMap}(e1, e2)
}
\arguments{
\item{e1}{an EarthEngineMap object.}

\item{e2}{an EarthEngineMap object.}
}
\description{
EarthEngineMap + EarthEngineMap; adds data from the second map to the first

EarthEngineMap | EarthEngineMap provides a slider in the middle to compare two maps.
}
\author{
tim-salabim. Adapted from mapview code.

tim-salabim. Adapted from mapview code.
}
