% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_operators.R
\name{+,EarthEngineMap,mapview-method}
\alias{+,EarthEngineMap,mapview-method}
\title{EarthEngineMap + ANY; adds data from the second map to the first}
\usage{
\S4method{+}{EarthEngineMap,mapview}(e1, e2)
}
\arguments{
\item{e1}{a EarthEngineMap map to which e2 should be added.}

\item{e2}{a EarthEngineMap map from which the objects should be added to e1.}
}
\description{
EarthEngineMap + ANY; adds data from the second map to the first
}
\author{
tim-salabim Adapted from mapview code.
}
