% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.R
\name{+,mapview,EarthEngineMap-method}
\alias{+,mapview,EarthEngineMap-method}
\title{ANY + EarthEngineMap; adds data from the second map to the first}
\usage{
\S4method{+}{mapview,EarthEngineMap}(e1, e2)
}
\arguments{
\item{e1}{a EarthEngineMap map to which e2 should be added.}

\item{e2}{a EarthEngineMap map from which the objects should be added to e1.}
}
\description{
ANY + EarthEngineMap; adds data from the second map to the first
}
\author{
Adapted from
\href{https://github.com/r-spatial/mapview/blob/develop/R/plus.R}{
tim-salabim} code.
}
