% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Map.R
\docType{methods}
\name{+,mapview,mapview-method}
\alias{+,mapview,mapview-method}
\title{mapview + mapview; adds data from the second map to the first}
\usage{
\S4method{+}{mapview,mapview}(e1, e2)
}
\arguments{
\item{e1}{a mapview map to which e2 should be added.}

\item{e2}{a mapview map from which the objects should be added to e1.}
}
\description{
mapview + mapview; adds data from the second map to the first
}
\examples{
\dontrun{
eeobject <- ee$FeatureCollection("users/csaybar/DLdemos/train_set")
center <- eeobject$geometry()$centroid()$getInfo()$coordinates
vizparams <- list(color = "FF0000", strokeWidth = 5)
m1 <- ee_map(eeobject, vizparams, center, objname = "Arequipa-landuse")

collection <- ee$ImageCollection("LANDSAT/LC08/C01/T1_TOA")$
  filter(ee$Filter()$eq("WRS_PATH", 44))$
  filter(ee$Filter()$eq("WRS_ROW", 34))$
  filterDate("2014-01-01", "2015-01-01")$
  sort("CLOUD_COVER")
eeobject <- collection$median()
vizparams <- list(bands = c("B4", "B3", "B2"), max = 0.3)
center <- c(-122.3578, 37.7726)
m2 <- ee_map(eeobject, vizparams, center, objname = "SF")
m1 + m2
}

}
\author{
Adapted from
\href{https://github.com/r-spatial/mapview/blob/develop/R/plus.R}{
tim-salabim code}.
}
