% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ee.computedobject.ComputedObject}
\alias{print.ee.computedobject.ComputedObject}
\alias{print}
\title{print Earth Engine object}
\usage{
\method{print}{ee.computedobject.ComputedObject}(x, ..., type = getOption("rgee.print.option"))
}
\arguments{
\item{x}{Earth Engine spatial object.}

\item{...}{ignored}

\item{type}{Character. What to show about the x object?. Three options are
supported: "json", "simply", "ee_print". By default "simply".}
}
\value{
No return value, called for displaying Earth Engine objects.
}
\description{
print Earth Engine object
}
