% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Date.R
\name{r_to_eeDate}
\alias{r_to_eeDate}
\title{Pass R date object to Earth Engine}
\usage{
r_to_eeDate(date)
}
\arguments{
\item{date}{R date object}
}
\description{
Pass R date object and return a timestamp as a float value
}
\examples{
ee_reattach()
ee_Initialize()
r_to_eeDate('2000-01-01')
}
