% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Date.R
\name{rdate_to_eedate}
\alias{rdate_to_eedate}
\title{Pass R date object to Earth Engine}
\usage{
rdate_to_eedate(date, eeobject = TRUE)
}
\arguments{
\item{date}{R date object}

\item{eeobject}{Logical. If TRUE, return
a ee$object otherwise return the date in milliseconds.}
}
\description{
Function to pass objects of classes "Date",
"character", "POSIXt", and "POSIXct" to
ee$Date or unix time stamp.
}
\examples{
library(rgee)
ee_reattach()
ee_Initialize()
rdate_to_eedate('2000-01-01')
rdate_to_eedate(315532800000) # float number
}
