% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_operators.R
\name{slider}
\alias{slider}
\alias{|}
\alias{|,}
\alias{EarthEngineMap,}
\alias{EarthEngineMap-method}
\alias{|.default}
\alias{|.EarthEngineMap}
\title{EarthEngineMap | EarthEngineMap provides a slider in the middle to compare two maps.}
\usage{
`|`(e1, e2)

\method{|}{default}(e1, e2)

\method{|}{EarthEngineMap}(e1, e2)
}
\arguments{
\item{e1}{an EarthEngineMap object.}

\item{e2}{an EarthEngineMap object.}
}
\description{
EarthEngineMap | EarthEngineMap provides a slider in the middle to compare two maps.
}
\author{
tim-salabim. Adapted from mapview code.
}
