% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_stars.R
\name{stars_as_ee}
\alias{stars_as_ee}
\title{Convert an sf object to an Earth Engine Image}
\usage{
stars_as_ee(x)
}
\arguments{
\item{x}{ee$Image}
}
\description{
Convert an sf object to an Earth Engine Image
}
