% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-math.R
\name{Summary-methods}
\alias{Summary-methods}
\alias{Summary.ee.image.Image}
\alias{mean.ee.image.Image}
\title{Summary Methods}
\usage{
\method{Summary}{ee.image.Image}(..., na.rm = TRUE)

\method{mean}{ee.image.Image}(..., na.rm = TRUE)
}
\arguments{
\item{...}{ee$Image.}

\item{na.rm}{Ignore.}
}
\description{
Summary Methods
}
