% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_geom.R
\name{ee_as_ee}
\alias{ee_as_ee}
\alias{ee_as_ee.data.frame}
\alias{ee_as_ee.character}
\alias{ee_as_ee.sf}
\alias{ee_as_ee.sfc}
\alias{ee_as_ee.sp}
\title{Convert foreign R spatial object to an Earth Engine object}
\usage{
ee_as_ee(x, ...)

\method{ee_as_ee}{data.frame}(x, coords, ...)

\method{ee_as_ee}{character}(x, ...)

\method{ee_as_ee}{sf}(x)

\method{ee_as_ee}{sfc}(x)

\method{ee_as_ee}{sp}(x)
}
\arguments{
\item{x}{objet to be converted into an Earth Engine object.}

\item{...}{additional \code{\link[sf]{st_as_sf.dataframe}} arguments could be passed.}

\item{coords}{in case of point data: names or numbers of the numeric columns holding coordinates}
}
\description{
Convert foreign R spatial object to an Earth Engine object
}
