% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_image.R
\name{ee_as_raster}
\alias{ee_as_raster}
\title{Convert an Earth Engine (EE) image into a raster object}
\usage{
ee_as_raster(
  image,
  region = NULL,
  dsn = NULL,
  via = "drive",
  scale = NULL,
  maxPixels = 1e+09,
  container = "rgee_backup",
  quiet = FALSE,
  ...
)
}
\arguments{
\item{image}{ee$Image to be converted into a raster object}

\item{region}{EE Geometry (ee$Geometry$Polygon) which specify the region
to export. CRS needs to be the same that the x argument otherwise it will be
forced. If not specified image bounds will be taken.}

\item{dsn}{Character. Output filename. If missing,
\code{ee_as_raster} will create a temporary file.}

\item{via}{Character. Method to fetch data about the object. Two methods
are implemented: "drive", "gcs". See details.}

\item{scale}{Numeric. The resolution in meters per pixel. Defaults
to the native resolution of the image asset.}

\item{maxPixels}{Numeric. The maximum allowed number of pixels in the
exported image. The task will fail if the exported region covers
more pixels in the specified projection. Defaults to 100,000,000.}

\item{container}{Character. Name of the folder ('drive') or bucket ('gcs')
to be exported into (ignored if \code{via} is not defined as "drive" or
"gcs").}

\item{quiet}{Logical. Suppress info message}

\item{...}{Extra exporting argument. See \link{ee_image_to_drive} and
\link{ee_image_to_gcs}.}
}
\value{
A RasterStack object
}
\description{
Convert an ee$Image into a raster object
}
\details{
\code{ee_as_raster} supports the download of \code{ee$Image}
by two different options: "drive" that use Google Drive and "gcs"
that use Google Cloud Storage. Previously, it is necessary to install the
R packages \href{ https://CRAN.R-project.org/package=googledrive}{googledrive}
or \href{https://CRAN.R-project.org/package=googleCloudStorageR}{
googleCloudStorageR} respectively. For getting more information about
exporting data from Earth Engine, take a look at the
\href{https://developers.google.com/earth-engine/exporting}{Google
Earth Engine Guide - Export data}.
}
\examples{
\dontrun{
library(rgee)

ee_Initialize(drive = TRUE, gcs = TRUE)
ee_user_info()

# Define an image.
img <- ee$Image("LANDSAT/LC08/C01/T1_SR/LC08_038029_20180810")$
  select(c("B4", "B3", "B2"))$
  divide(10000)

# OPTIONAL display it using Map
Map$centerObject(eeObject = img)
Map$addLayer(eeObject = img, visParams = list(max = 0.4,gamma=0.1))

# Define an area of interest.
geometry <- ee$Geometry$Rectangle(
  coords = c(-110.8, 44.6, -110.6, 44.7),
  proj = "EPSG:4326",
  geodesic = FALSE
)

## drive - Method 01
img_02 <- ee_as_raster(
  image = img,
  region = geometry,
  via = "drive"
)

## gcs - Method 02
# img_03 <- ee_as_raster(
#   image = img,
#   region = geometry,
#   container = "rgee_dev",
#   via = "gcs"
# )

# OPTIONAL: Delete containers
# ee_clean_container(name = "rgee_backup", type = "drive")
# ee_clean_container(name = "rgee_dev", type = "gcs")
}
}
\seealso{
Other image download functions: 
\code{\link{ee_as_stars}()},
\code{\link{ee_as_thumbnail}()},
\code{\link{ee_imagecollection_to_local}()}
}
\concept{image download functions}
