% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_stars.R
\name{ee_as_stars}
\alias{ee_as_stars}
\title{Convert an Earth Engine (EE) image in a stars object}
\usage{
ee_as_stars(image, region, scale = NULL, via = "getInfo",
  container = "rgee_backup", quiet = FALSE)
}
\arguments{
\item{image}{ee$Image to be converted into a sf object}

\item{region}{ee$Geometry$Polygon. Region of interest}

\item{scale}{The resolution in meters per pixel. If scale
is NULL, the native resolution of the first band will be
taken.}

\item{via}{Method to download the image. Three methods
are implemented 'getInfo', 'drive' and 'gcs'. See details.}

\item{container}{Relevant when the "via" argument is
defined as 'drive' or 'gcs'. The name of a unique
folder ('drive') or bucket ('gcs') to export into.}

\item{quiet}{logical. Suppress info message}
}
\value{
A stars object, when via argument was defined as 'getInfo'. For
other cases will return a stars-proxy object.
}
\description{
Convert an Earth Engine (EE) image in a stars object
}
\details{
The process to pass a ee$Image to your local env could be carried
out by three different strategies. The first one ('getInfo') use the getInfo
image method, which fetch and return information about Earth Engine
images, the advantage of use this is a direct and fast download.
However, there is a limit of 262144 pixels (512x512) that can
be transferred by request which makes it unsatisfactory for large images.
The second ('drive') and third ('gcs') method are suitable for large images
since it use Google Drive and Google Cloud Storage as a intermidiate
container.
}
\examples{
library(raster)
library(stars)
library(rgee)
# Initialize a specific Earth Engine account and
# load Google Drive and Google Cloud Storage credentials
ee_reattach()
ee_Initialize(
  email = "data.colec.fbf@gmail.com",
  drive = TRUE,
  gcs = TRUE
)

# Define an image.
img <- ee$Image('LANDSAT/LC08/C01/T1_SR/LC08_038029_20180810')$
  select(c('B4', 'B5', 'B6'))
Map$centerObject(img)
Map$addLayer(img)

# Define an area of interest.
geometry <- ee$Geometry$Rectangle(c(-110.8, 44.6, -110.6, 44.7))

## getInfo - Method 01
img_stars_01 <- ee_as_stars(image = img,
                            region = geometry,
                            via = 'getInfo')
## drive - Method 02
img_stars_02 <- ee_as_stars(image = img,
                            region = geometry,
                            via = 'drive')
## gcs - Method 03
img_stars_03 <- ee_as_stars(image = img,
                            region = geometry,
                            container = 'rgee_dev',
                            via = 'gcs')
}
