% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_stars.R
\name{ee_as_stars}
\alias{ee_as_stars}
\title{Convert an Earth Engine (EE) image in a stars object}
\usage{
ee_as_stars(image, region, via = "getInfo", quiet = FALSE)
}
\arguments{
\item{image}{ee$Image to be converted into a sf object}

\item{region}{ee$Geometry$Polygon. Region of interest}

\item{quiet}{logical. Suppress info message}
}
\description{
Convert an Earth Engine (EE) image in a stars object
}
\examples{
library(raster)
library(stars)
library(rgee)
# Define an image.
img <- ee$Image('LANDSAT/LC08/C01/T1_SR/LC08_038029_20180810')$
          select(c('B4', 'B5', 'B6'))
# Define an area of interest.
geometry <- ee$Geometry$Rectangle(c(-110.8, 44.6, -110.6, 44.7))

# Download Image an area of interest.
# ee_help(img$sampleRectangle)
img_stars <- ee_as_stars(image = img,region = geometry)
image(img_stars, rgb = c(3,2,1))
as(img_stars,'Raster') \%>\% plotRGB(3,2,1)
}
