% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_check.R
\name{ee_check-tools}
\alias{ee_check-tools}
\alias{ee_check}
\alias{ee_check_python}
\alias{ee_check_python_modules}
\alias{ee_check_drivers}
\alias{ee_check_credentials}
\alias{ee_install_drivers}
\alias{ee_get_earthengine_path}
\title{Interface to handle Installation requirements}
\usage{
ee_check()

ee_check_python()

ee_check_python_modules()

ee_check_drivers()

ee_check_credentials()

ee_install_drivers(version = "latest")

ee_get_earthengine_path()
}
\description{
R functions for handle: drivers (geckodrivers), credentials and python packages.
}
