% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_stars.R
\name{ee_clean_container}
\alias{ee_clean_container}
\title{Delete files from a file container}
\usage{
ee_clean_container(name = "rgee_backup", type = "drive")
}
\arguments{
\item{name}{Character. Name of the folder (Drive) or bucket (GCS)
to delete all files into.}

\item{type}{Character. Name of the file storage web service. 'drive'
and 'gcs' are supported.}
}
\description{
Delete all files from a folder (Google Drive) or a bucket
(Google Cloud Storage). Caution: This will permanently delete
your backup files generated by using ee_as_stars and ee_as_sf.
}
