% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_remove.R
\name{ee_clean_pyenv}
\alias{ee_clean_pyenv}
\title{Remove reticulate system variables from .Renviron}
\usage{
ee_clean_pyenv()
}
\description{
Remove reticulate system variables from .Renviron
}
\examples{
\dontrun{
library(rgee)
ee_clean_pyenv()
}
}
