% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_clean.R
\name{ee_clean_pyenv}
\alias{ee_clean_pyenv}
\title{Remove rgee system variables from .Renviron}
\usage{
ee_clean_pyenv(home)
}
\value{
No return value, called for cleaning environmental variables in their system.
}
\description{
Remove rgee system variables from .Renviron
}
\seealso{
Other ee_clean functions: 
\code{\link{ee_clean_container}()},
\code{\link{ee_clean_credentials}()}
}
\concept{ee_clean functions}
