% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_discover_pyenvs}
\alias{ee_discover_pyenvs}
\title{Discover all the Python environments available
in the system}
\usage{
ee_discover_pyenvs(use_py_discover_config = FALSE)
}
\arguments{
\item{use_py_discover_config}{Logical. If TRUE
will use
\code{reticulate::\link[reticulate:py_discover_config]{py_discover_config}}
to find versions of Python in the system.  Otherwise, will use
\code{reticulate::\link[reticulate:conda_list]{conda_list}}
for Window OS and
\code{reticulate::\link[reticulate:virtualenv_list]{virtualenv_list}}
for Unix system.}
}
\description{
This function enables callers to check which versions
of Python will be discovered on a system.
}
\examples{
\dontrun{
library(rgee)
ee_create_pyenv('ee')
ee_discover_pyenvs()
ee_set_pyenv(
  python_path = ".../ee/bin/python",
  python_env = 'ee')
ee_install_python_packages()
}
}
