% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_aux.R
\name{ee_Download}
\alias{ee_Download}
\title{Soy una bonita que aun no hace nada}
\usage{
ee_Download(task, path = NULL, overwrite = FALSE)
}
\arguments{
\item{task}{holi2}

\item{path}{holi2}

\item{overwrite}{holi2}
}
\description{
Soy una bonita que aun no hace nada
}
