% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_extract.R
\name{ee_extract}
\alias{ee_extract}
\title{Extract values from Image EE spatial objects
Extract values from a Image or ImageCollection spatial object at the locations
of geometry object. You can use coordinates, ee.Geometries, ee.Features,
ee.FeatureCollection and sf objects.}
\usage{
ee_extract(x, y, fun = ee$Reducer$mean(), scale = 1000, sf = TRUE,
  ...)
}
\arguments{
\item{x}{ee$Image or ee$ImageCollection.}

\item{y}{ee$Geometry, ee$Feature, ee$FeatureCollection or sf object.}

\item{fun}{ee$Reducer object. Function to summarize the values. See details.}

\item{scale}{TODO}

\item{sf}{TODO}

\item{...}{reduceRegions parameters. See Details.
library(rgee)
ee_Initialize()
terraclimate = ee$ImageCollection("IDAHO_EPSCOR/TERRACLIMATE")$
filterDate("2000-01-01","2001-01-01")$
map(function(x) x$select("pr"))

ee_nc = st_read(system.file("shape/nc.shp", package="sf")) %>%
st_transform(4326) %>%
sf_as_ee()
ee_nc_rain = ee_extract(terraclimate,ee_nc)
ee_nc_rain}
}
\description{
Extract values from Image EE spatial objects
Extract values from a Image or ImageCollection spatial object at the locations
of geometry object. You can use coordinates, ee.Geometries, ee.Features,
ee.FeatureCollection and sf objects.
}
