% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_functions.R
\name{ee_FeatureCollection}
\alias{ee_FeatureCollection}
\title{Constructs an Earth Engine FeatureCollection}
\usage{
ee_FeatureCollection(args, opt_column = NULL)
}
\arguments{
\item{args}{constructor argument. One of:
\enumerate {
\item A string - assumed to be the name of a collection.
\item A geometry.
\item A feature.
\item An array of features.
\item A computed object - reinterpreted as a collection.
}}

\item{opt_column}{The name of the geometry column to use. Only useful with the
string constructor.}
}
\description{
Constructs an Earth Engine FeatureCollection
}
