% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Initialize.R
\name{ee_get_assethome}
\alias{ee_get_assethome}
\title{Get the Asset home name}
\usage{
ee_get_assethome()
}
\description{
Get the Asset home name
}
\examples{
\dontrun{
library(rgee)
ee_reattach()
ee_Initialize()
ee_get_assethome()
}
}
