% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Date.R
\name{ee_get_date_ic}
\alias{ee_get_date_ic}
\title{Get the date of a EE ImageCollection}
\usage{
ee_get_date_ic(x, time_end = FALSE)
}
\arguments{
\item{x}{EE ImageCollection object}

\item{time_end}{Logical. If TRUE, the system:time_end property will
also be returned. See details.}
}
\value{
A data.frame.
}
\description{
Get the date of a EE ImageCollection
}
\details{
\code{system:time_start} set the start period of data acquisition while
\code{system:time_end} the end period. See the
\href{https://developers.google.com/earth-engine/glossary}{Earth
Engine glossary} for more information.
}
\examples{
\dontrun{
library(rgee)
library(sf)

ee_reattach()
ee_Initialize()

nc <- st_read(system.file("shape/nc.shp", package = "sf")) \%>\%
  st_transform(4326) \%>\%
  sf_as_ee()

ee_s2 <- ee$ImageCollection("COPERNICUS/S2")$
  filterDate("2016-01-01", "2016-01-31")$
  filterBounds(nc)

ee_get_date_ic(ee_s2)
}
}
