% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Initialize.R
\name{ee_get_earthengine_path}
\alias{ee_get_earthengine_path}
\title{Get the path where rgee save the credentials}
\usage{
ee_get_earthengine_path()
}
\description{
Get the path where rgee save the credentials
}
