% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_functions.R
\name{ee_ImageCollection}
\alias{ee_ImageCollection}
\title{Representation for an Earth Engine ImageCollection}
\usage{
ee_ImageCollection(args)
}
\arguments{
\item{args}{ImageCollections can be constructed from the following arguments:
\enumerate{
\item A string: assumed to be the name of a collection,
\item An array of images, or anything that can be used to construct an
image.
\item A single image.
\item A computed object - reinterpreted as a collection.
}}
}
\description{
Representation for an Earth Engine ImageCollection
}
