% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_imagecollection.R
\name{ee_imagecollection_to_local}
\alias{ee_imagecollection_to_local}
\title{Save an EE ImageCollection in their local system}
\usage{
ee_imagecollection_to_local(
  ic,
  region,
  dsn = NULL,
  via = "getInfo",
  scale = NULL,
  maxPixels = 1e+09,
  container = "rgee_backup",
  quiet = FALSE,
  ...
)
}
\arguments{
\item{ic}{ee$ImageCollection to be saved in the system.}

\item{region}{EE Geometry Rectangle (ee$Geometry$Rectangle). The
CRS needs to be the same that the ic argument otherwise it will be
forced.}

\item{dsn}{Character. Output filename. If missing,
\code{ee_imagecollection_to_local} will create a temporary file.}

\item{via}{Character. Method to fetch data about the object. Multiple
options supported. See details.}

\item{scale}{Numeric. The resolution in meters per pixel. Defaults
to the native resolution of the image assset.}

\item{maxPixels}{Numeric. The maximum allowed number of pixels in the
exported image. The task will fail if the exported region covers
more pixels in the specified projection. Defaults to 100,000,000.}

\item{container}{Character. Name of the folder ('drive') or bucket ('gcs')
to be exported into (ignored if \code{via} is not defined as "drive" or
"gcs").}

\item{quiet}{logical. Suppress info message}

\item{...}{Extra exporting argument. See \link{ee_image_to_drive} and
\link{ee_image_to_gcs}.}
}
\value{
Character vector containing the filename of the images downloaded.
}
\description{
Save an EE ImageCollection in their local system
}
\details{
\code{ee_imagecollection_to_local} supports the download of \code{ee$Image}
by three different options: "getInfo", "drive", and "gcs". When "getInfo"
is set in the \code{via} argument, \code{ee_imagecollection_to_local} will
make an REST call to retrieve all the known information about the object.
The advantage of use "getInfo" is a direct and faster download. However,
there is a limitation of 262144 pixels by request which makes it not
recommendable for large images. Instead of "getInfo", the options: "drive"
and "gcs" are suitable for large collections since they use an intermediate
container. They use Google Drive and Google Cloud Storage respectively. For
getting more information about exporting data from Earth Engine,  take a
look at the \href{https://developers.google.com/earth-engine/exporting}{Google
Earth Engine Guide - Export data}.
}
\examples{
\dontrun{
library(rgee)
library(raster)
ee_Initialize(drive = TRUE, gcs = TRUE)

# USDA example
loc <- ee$Geometry$Point(-99.2222, 46.7816)
collection <- ee$ImageCollection('USDA/NAIP/DOQQ')$
  filterBounds(loc)$
  filterDate('2008-01-01', '2020-01-01')$
  filter(ee$Filter$listContains("system:band_names", "N"))

# From ImageCollection to local directory
ee_crs <- collection$first()$projection()$getInfo()$crs
geometry <- collection$first()$geometry(proj = ee_crs)$bounds()
tmp <- tempdir()

## Using drive
ic_drive_files <- ee_imagecollection_to_local(
  ic = collection,
  region = geometry,
  scale = 100,
  dsn = file.path(tmp, "drive_"),
  via = "drive"
)

}
}
\seealso{
Other image download functions: 
\code{\link{ee_as_raster}()},
\code{\link{ee_as_stars}()},
\code{\link{ee_as_thumbnail}()}
}
\concept{image download functions}
