% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{ee_initialize}
\alias{ee_initialize}
\title{Authenticate and Initialize the Earth Engine library.}
\usage{
ee_initialize(get_credentials = FALSE)
}
\arguments{
\item{get_credentials}{TODO}
}
\description{
R wrapper for ee.Initialize(): Authenticate and Initialize the Earth Engine API. See details.
}
\details{
The \code{ee_initialize()} will allow you, via web-browser, obtain the
\href{https://developers.google.com/earth-engine/python_install_manual#setting-up-authentication-credentials}{OAuth 2.0 Client Credential}
. It is necessary for getting access to Google Earth Engine. Copy and
paste the key into the R console when prompted for the key.
}
\examples{
\dontrun{
ee_initialize()
}
}
