% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{ee_initialize}
\alias{ee_initialize}
\title{Initialize the Earth Engine library.}
\usage{
ee_initialize(get_credentials = FALSE)
}
\arguments{
\item{get_credentials}{TODO}
}
\description{
Earth Engine uses the OAuth 2.0 protocol for authenticating clients.
Running \code{ee_initialize} will prompt you through the authentication process
using your web browser. See Setting Up Authentication Credentials in the
\href{https://developers.google.com/earth-engine/python_install_manual}{Python API documentation}
for additional details.
}
