% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_auth.R
\name{ee_initialize}
\alias{ee_initialize}
\title{Authenticate and Initialize the Earth Engine and Google Drive API}
\usage{
ee_initialize()
}
\description{
Authenticate and Initialize the Earth Engine and Google Drive API
}
\details{
The \code{ee_initialize()} via web-browser asked to sign in to your Google account and
grant permission for both google earth engine and google drive. The user credentials
are save in the folder \code{~/.config/earthengine}.
}
\examples{
\dontrun{
ee_initialize()
}
}
