% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_install-tools}
\alias{ee_install-tools}
\alias{ee_install_drivers}
\alias{ee_install_rgee_python_packages}
\alias{ee_install_python_package}
\alias{ee_install_python_ee}
\alias{ee_remove_drivers}
\alias{ee_remove_credentials}
\title{Interface to install non-R rgee dependencies}
\usage{
ee_install_drivers(version)

ee_install_rgee_python_packages(conda = FALSE)

ee_install_python_package(pypackage, conda = FALSE)

ee_install_python_ee(conda = FALSE)

ee_remove_drivers(quiet = TRUE)

ee_remove_credentials(user, quiet = TRUE)
}
\arguments{
\item{version}{TODO}

\item{conda}{TODO}

\item{pypackage}{TODO}

\item{quiet}{logical; suppress info message}

\item{user}{TODO}
}
\description{
R functions for installing and removing Python Selenium Chromedriver and Third-Party Python packages.
}
\details{
It is neccessary restart RStudio to observe change when installing a python packages. rgee only is
compatible with Python 3.5 >.
}
\examples{
\dontrun{
library(rgee)

# Recommended way to install external dependencies of rgee for Linux users
virtualenv_remove("rgee")
virtualenv_create("rgee", python = "python3.5")
use_virtualenv("rgee")

ee_check()
ee_install_rgee_python_packages() # Install rgee python packages
ee_install_drivers() # Install selenim drivers
ee_get_credentials() # Get credentials
ee_check()

# Recommended way to install external dependencies of rgee for Windows users
}
}
