% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_install}
\alias{ee_install}
\title{Install rgee dependecies}
\usage{
ee_install(packages, method = c("auto", "virtualenv", "conda"),
  conda = "auto")
}
\arguments{
\item{packages}{Character vector with package names to install.}

\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows.}

\item{conda}{conda Path to conda executable (or "auto" to find conda using
the PATH and other conventional install locations).}
}
\description{
Wrapping to reticulate::py_install(packages = "ee")
}
\details{
On Linux and OS X the "virtualenv" method will be used by default
("conda" will be used if virtualenv isn't available). On Windows, the
"conda" method is always used.
}
\author{
\href{https://github.com/kevinushey}{Kevin Ushey} and \href{https://github.com/jjallaire}{J.J. Allaire}
}
