% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_install_create_pyenv}
\alias{ee_install_create_pyenv}
\title{Create an isolated Python virtual environment to be used in rgee}
\usage{
ee_install_create_pyenv(py_env = "rgee")
}
\arguments{
\item{py_env}{The name of, or path to, a Python virtual environment.}
}
\value{
Character. The path of the virtual environment created.
}
\description{
Create an isolated Python virtual environment to be used in rgee
}
\examples{
\dontrun{
library(rgee)

## It is necessary just once, not mandatory

# 1. Create a Python environment, e.g. ee.
pyenv <- ee_install_create_pyenv(py_env = "ee")

# OPTIONAL: Find others Python path in the system.
# ee_install_discover_pyenvs()

# 2. Set a Python path in .Renviron (EARTHENGINE_PYTHON)
# to be used in future sessions
ee_install_set_pyenv(pyenv)

# 3. Now run ee_Initialize()
ee_Initialize()
}
}
