% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_install_discover_pyenvs}
\alias{ee_install_discover_pyenvs}
\title{Discover all the Python environments available
in the system}
\usage{
ee_install_discover_pyenvs(use_py_discover_config = FALSE)
}
\arguments{
\item{use_py_discover_config}{Logical. If TRUE
will use \code{\link{reticulate}} to find
versions of Python in the system.  Otherwise, will use
\link[=reticulate]{conda_list} for Window OS and
\link[=reticulate]{virtualenv_list} for Unix system.}
}
\value{
Python configuration object (reticulate).
}
\description{
This function enables callers to check which versions
of Python will be discovered on a system.
}
\examples{
\dontrun{
#' library(rgee)

### rgee installation

# 1. Initialize rgee with ee_Initialize(). If there is no any Python
# environment, miniconda will be installed by default.
ee_Initialize()

# 2. Create a Python environment, e.g. ee.
pyenv <- ee_install_create_pyenv(py_env = "ee")

# Find others Python environments in the system.
ee_install_discover_pyenvs()

# 3. Set a Python environment (e.g. ee) and restart R to see changes.
ee_install_set_pyenv(pyenv, install = TRUE)

# 4. Install Python package dependencies and restart R to see changes.
ee_install_python_packages()

# 5. Initialize rgee again!
ee_Initialize()
}
}
