% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_install_set_pyenv}
\alias{ee_install_set_pyenv}
\title{Set the Python environment to be used on rgee}
\usage{
ee_install_set_pyenv(py_path)
}
\arguments{
\item{py_path}{The path to a Python interpreter, to be used with rgee.}
}
\description{
Set the Python environment to be used on rgee
}
