% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_install_set_pyenv}
\alias{ee_install_set_pyenv}
\title{Set the Python environment to be used by rgee}
\usage{
ee_install_set_pyenv(py_path, py_env = "rgee")
}
\arguments{
\item{py_path}{The path to a Python interpreter.}

\item{py_env}{The name of the environment. By default "rgee".}
}
\description{
This function create a new environment variable called 'EARTHENGINE_PYTHON'.
It is used to set the Python environment to be used by rgee.
EARTHENGINE_PYTHON is saved into the file .Renviron.
}
\seealso{
Other ee_install functions: 
\code{\link{ee_install}()}
}
\concept{ee_install functions}
