% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_install_set_pyenv}
\alias{ee_install_set_pyenv}
\title{Set the Python environment to be used on rgee}
\usage{
ee_install_set_pyenv(py_path)
}
\arguments{
\item{py_path}{The path to a Python interpreter, to be used with rgee.}
}
\description{
Set the Python environment to be used on rgee
}
\examples{
\dontrun{
library(rgee)

## It is necessary just once, not mandatory

# 1. Create a Python environment, e.g. ee.
pyenv <- ee_install_create_pyenv(py_env = "ee")

# OPTIONAL: Find others Python path in the system.
# ee_install_discover_pyenvs()

# 2. Set a Python path in .Renviron (EARTHENGINE_PYTHON)
# to be used in future sessions
ee_install_set_pyenv(pyenv)

# 3. Now run ee_Initialize()
ee_Initialize()
}
}
