% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_install_set_pyenv}
\alias{ee_install_set_pyenv}
\title{Set the Python environment to be used by rgee}
\usage{
ee_install_set_pyenv(py_path = NULL, py_env = NULL, quiet = FALSE)
}
\arguments{
\item{py_path}{The path to a Python interpreter}

\item{py_env}{The name of the environment}

\item{quiet}{Logical. Suppress info message}
}
\value{
no return value, called for setting EARTHENGINE_PYTHON in .Renviron
}
\description{
This function create a new environment variable called 'EARTHENGINE_PYTHON'.
It is used to set the Python environment to be used by rgee.
EARTHENGINE_PYTHON is saved into the file .Renviron.
}
\seealso{
Other ee_install functions: 
\code{\link{ee_install_upgrade}()},
\code{\link{ee_install}()}
}
\concept{ee_install functions}
