% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_install.R
\name{ee_install_upgrade}
\alias{ee_install_upgrade}
\title{Upgrade the Earth Engine Python API}
\usage{
ee_install_upgrade(
  version = NULL,
  earthengine_env = Sys.getenv("EARTHENGINE_ENV")
)
}
\arguments{
\item{version}{Character. The Earth Engine Python API version to upgrade.
By default \code{rgee::ee_version()}.}

\item{earthengine_env}{Character. The name, or full path, of the
environment in which the earthengine-api packages are to be installed.}
}
\value{
no return value, called to upgrade the earthengine-api Python package
}
\description{
Upgrade the Earth Engine Python API
}
\examples{
\dontrun{
library(rgee)
# ee_install_upgrade()
}
}
\seealso{
Other ee_install functions: 
\code{\link{ee_install}()},
\code{\link{ee_install_set_pyenv}()}
}
\concept{ee_install functions}
