% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_manage.R
\name{ee_manage-tools}
\alias{ee_manage-tools}
\alias{ee_manage_create}
\alias{ee_manage_delete}
\alias{ee_manage_assetlist}
\alias{ee_manage_quota}
\alias{ee_manage_size}
\alias{ee_manage_copy}
\alias{ee_manage_move}
\alias{ee_manage_set_properties}
\alias{ee_manage_delete_properties}
\alias{ee_manage_assets_access}
\alias{ee_manage_task}
\alias{ee_manage_cancel_all_running_taks}
\title{Interface for manage the Earth Engine Asset}
\usage{
ee_manage_create(path_asset, asset_type = "folder", quiet = FALSE)

ee_manage_delete(path_asset, quiet = FALSE)

ee_manage_assetlist(path_asset, quiet = FALSE)

ee_manage_quota()

ee_manage_size(path_asset)

ee_manage_copy(path_asset, final_path, quiet = FALSE)

ee_manage_move(path_asset, final_path, quiet = FALSE)

ee_manage_set_properties(path_asset, properties)

ee_manage_delete_properties(path_asset, property)

ee_manage_assets_access(path_asset,
  acl = getOption("rgee.manage.getAssetAcl"), quiet = FALSE)

ee_manage_task(quiet, cache = TRUE)

ee_manage_cancel_all_running_taks()
}
\arguments{
\item{path_asset}{Character. Name of the EE asset e.g. users/user/nameofcollection.}

\item{asset_type}{Character. The asset type ('folder' or 'imagecollection').}

\item{quiet}{Logical. Suppress info message.}

\item{final_path}{Character. Name of the final EE asset to copy or move.}

\item{properties}{List. Set of parameters to established as a property of an EE object. See details.}

\item{property}{Character. Name of a specific property to be deleted.}

\item{acl}{A list describing the asset's Access Control List. See getOption("rgee.manage.getAssetAcl").}

\item{cache}{Logical. TRUE save the report in the /temp directory.}
}
\description{
R functions for managing the Earth Engine Asset
}
\examples{
library(rgee)
# Change the user and google account to be able to reproduce
ee_Initialize(user_gmail = 'csaybar@gmail.com')
ee_manage_create('users/csaybar/rgee')

# 1. List all the elements inside a folder or a ImageCollection
ee_manage_assetlist(path_asset = 'users/csaybar/rgee')

# 2. Create a Folder or a ImageCollection
ee_manage_create('users/csaybar/rgee/rgee_folder',asset_type = 'folder')
ee_manage_create('users/csaybar/rgee/rgee_ic',asset_type = 'imagecollection')
ee_manage_assetlist('users/csaybar/rgee')

# 3. Shows your Earth Engine quota
ee_manage_quota()

# 4. Estimate the size of a Image, ImageCollection, Table or Folder.
ee_manage_size('users/csaybar/rgee')

# 5. Move a EE object to another folder
ee_manage_move(path_asset = 'users/csaybar/rgee/rgee_ic',
               final_path = 'users/csaybar/rgee/rgee_folder/rgee_ic_moved')
ee_manage_assetlist('users/csaybar/rgee/rgee_folder')

# 6. Set properties to an EE object.
ee_manage_set_properties(path_asset = 'users/csaybar/rgee/rgee_folder/rgee_ic_moved',
                         properties = list(message='hello-world',language = 'R'))
test_ic <- ee$ImageCollection('users/csaybar/rgee/rgee_folder/rgee_ic_moved')
test_ic$getInfo()

# 7. Delete properties
ee_manage_delete_properties(path_asset = 'users/csaybar/rgee/rgee_folder/rgee_ic_moved',
                            property = c("message","language"))
test_ic$getInfo()

# 8. Share EE objects -- Create a public dataset
ee_manage_assets_access('users/csaybar/rgee/rgee_folder/rgee_ic_moved')
ee$data$getAssetAcl('users/csaybar/rgee/rgee_folder/rgee_ic_moved')

# 9. Create a report based on all tasks that is running or has finished
ee_manage_task()

# 10. Cancel all the running task
ee_manage_cancel_all_running_taks()

# 11. Delete EE objects or folders
ee_manage_delete('users/csaybar/rgee/')
}
