% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_manage.R
\name{ee_manage-tools}
\alias{ee_manage-tools}
\alias{ee_manage_create}
\alias{ee_manage_delete}
\alias{ee_manage_assetlist}
\alias{ee_manage_quota}
\alias{ee_manage_copy}
\alias{ee_manage_move}
\alias{ee_manage_set_properties}
\alias{ee_manage_delete_properties}
\alias{ee_manage_task}
\alias{ee_manage_cancel_all_running_taks}
\title{Interface for manage the Earth Engine Asset}
\usage{
ee_manage_create(path_asset, asset_type = "Folder", quiet = FALSE)

ee_manage_delete(path_asset, quiet = FALSE)

ee_manage_assetlist(path_asset, quiet = FALSE)

ee_manage_quota()

ee_manage_copy(path_asset, final_path, quiet = FALSE)

ee_manage_move(path_asset, final_path, quiet = FALSE)

ee_manage_set_properties(path_asset, add_properties)

ee_manage_delete_properties(path_asset, del_properties = "ALL")

ee_manage_task(cache = TRUE)

ee_manage_cancel_all_running_taks()
}
\arguments{
\item{path_asset}{Character. Name of the EE asset e.g. users/user/nameofcollection.}

\item{asset_type}{Character. The asset type ('Folder' or 'ImageCollection').}

\item{quiet}{Logical. Suppress info message.}

\item{final_path}{Character. Name of the final EE asset to copy or move.}

\item{add_properties}{List. Set of parameters to established as a property of an EE object. See details.}

\item{del_properties}{Character. Name of a specific property to be deleted.}

\item{cache}{Logical. TRUE save the report in the /temp directory.}
}
\description{
R functions for managing the Earth Engine Asset
}
\examples{
\dontrun{
library(rgee)
ee_Initialize()

# Change google account to be able to reproduce
ee_manage_create("users/datacolecfbf/rgee")

# 1. List all the elements inside a folder or a ImageCollection
ee_manage_assetlist(path_asset = "users/datacolecfbf/rgee")

# 2. Create a Folder or a ImageCollection
ee_manage_create("users/datacolecfbf/rgee/rgee_folder", asset_type = "FOLDER")
ee_manage_create("users/datacolecfbf/rgee/rgee_ic", asset_type = "ImageCollection")
ee_manage_assetlist("users/datacolecfbf/rgee")

# 3. Shows your Earth Engine quota
# v ee_manage_quota()

# 4. Estimate the size of a IMAGE, ImageCollection, Table or folder.
ee_manage_size("users/datacolecfbf/rgee")

# 5. Move a EE object to another folder
ee_manage_move(
  path_asset = "users/datacolecfbf/rgee/rgee_ic",
  final_path = "users/datacolecfbf/rgee/rgee_folder/rgee_ic_moved"
)
ee_manage_assetlist("users/datacolecfbf/rgee/rgee_folder")

# 6. Set properties to an EE object.
ee_manage_set_properties(
  path_asset = "users/datacolecfbf/rgee/rgee_folder/rgee_ic_moved",
  properties = list(message = "hello-world", language = "R")
)
test_ic <- ee$ImageCollection("users/datacolecfbf/rgee/rgee_folder/rgee_ic_moved")
test_ic$getInfo()

# 7. Delete properties
ee_manage_delete_properties(
  path_asset = "users/datacolecfbf/rgee/rgee_folder/rgee_ic_moved",
  property = c("message", "language")
)
test_ic$getInfo()

# 8. Share EE objects -- Create a public dataset
ee_manage_assets_access("users/datacolecfbf/rgee/rgee_folder/rgee_ic_moved")
ee$data$getAssetAcl("users/datacolecfbf/rgee/rgee_folder/rgee_ic_moved")

# 9. Create a report based on all tasks that is running or has finished
ee_manage_task()

# 10. Cancel all the running task
ee_manage_cancel_all_running_taks()

# 11. Delete EE objects or folders
ee_manage_delete("users/datacolecfbf/rgee/")
}
}
