% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_manage.R
\name{ee_manage-tools}
\alias{ee_manage-tools}
\alias{ee_manage_create}
\alias{ee_manage_delete}
\alias{ee_manage_assetlist}
\alias{ee_manage_quota}
\alias{ee_manage_size}
\alias{ee_manage_copy}
\alias{ee_manage_move}
\alias{ee_manage_set_properties}
\alias{ee_manage_delete_properties}
\alias{ee_manage_assets_access}
\alias{ee_manage_task}
\alias{ee_manage_cancel_all_running_taks}
\title{Create a folder or ImageCollection into GEE assets}
\usage{
ee_manage_create(path_asset, asset_type = "folder", quiet = FALSE)

ee_manage_delete(path_asset, quiet = FALSE)

ee_manage_assetlist(path_asset, quiet = FALSE)

ee_manage_quota()

ee_manage_size(path_asset)

ee_manage_copy(path_asset, final_path, quiet = FALSE)

ee_manage_move(path_asset, final_path, quiet = FALSE)

ee_manage_set_properties(path_asset, properties)

ee_manage_delete_properties(path_asset, property)

ee_manage_assets_access(path_asset,
  acl = getOption("rgee.manage.getAssetAcl"), quiet = FALSE)

ee_manage_task(quiet)

ee_manage_cancel_all_running_taks()
}
\arguments{
\item{path_asset}{a character vector containing a single path name.}

\item{asset_type}{a character vector containing the asset type. 'folder' or 'imagecollection'.}

\item{quiet}{logical; suppress info message.}

\item{final_path}{TODO}

\item{properties}{TODO}

\item{property}{TODO}

\item{acl}{TODO}
}
\description{
Create a folder or ImageCollection into GEE assets
}
