% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_map.R
\name{ee_map}
\alias{ee_map}
\title{Display a given EE object}
\usage{
ee_map(eeobject, vizparams, center, zoom_start = 8, max_nimage = 10,
  obj_name = "map", ...)
}
\arguments{
\item{eeobject}{ee_Feature,ee_FeatureCollection,ee_Image or ee_ImageCollection;
the object to add to the map.}

\item{vizparams}{List; The visualization parameters. See Details}

\item{center}{Character vector; the longitude and latitude of the map center}

\item{zoom_start}{The zoom level.}

\item{max_nimage}{Specify the max number of Image to display (parameter appropriate just for ee_ImageCollection objects).}

\item{obj_name}{Character vector; Name of the map(or maps).}

\item{...}{additional \link[mapview]{mapview} arguments could be passed.}
}
\description{
Display a given EE object
}
\examples{
\dontrun{
library(rgee)
ee_initialize()

#Ploting an ee_FeatureColletion
eeobject <- ee_FeatureCollection('users/csaybar/DLdemos/train_set')
center <- eeobject$geometry()$centroid()$getInfo()$coordinates
vizparams = list(color='FF0000', strokeWidth=5)
ee_map(eeobject,vizparams,center,obj_name = "Arequipa-landuse")

#Ploting an ee_Image
collection = ee_ImageCollection('LANDSAT/LC08/C01/T1_TOA')$
 filter(ee_Filter()$eq('WRS_PATH', 44))$
 filter(ee_Filter()$eq('WRS_ROW', 34))$
 filterDate('2014-01-01', '2015-01-01')$
 sort('CLOUD_COVER')

eeobject = collection$median()
vizparams = list(bands=c('B4', 'B3', 'B2'),max=0.3)
center = c(-122.3578,37.7726)
ee_map(eeobject,vizparams,center,obj_name = 'SF')

#Ploting an ee_ImageColection
ee_map(collection,vizparams,center,obj_name = c('SF1',"SF2","SF3"),max_nimage = 3)
}
}
