% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_download.R
\name{ee_monitoring}
\alias{ee_monitoring}
\title{Monitoring Earth Engine task progress}
\usage{
ee_monitoring(task, eeTaskList = FALSE)
}
\arguments{
\item{task}{List generated after finished correctly a EE task, if it is missing the last one will be taken.}

\item{eeTaskList}{Logical, if \code{TRUE}, all Earth Engine tasks will listing initially.}
}
\description{
Monitoring Earth Engine task progress
}
\examples{
\dontrun{
library(rgee)
ee_initialize()
ee_download_monitoring(eelist = TRUE)
}
}
