% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_pyfunc.R
\name{ee_pyfunc}
\alias{ee_pyfunc}
\title{Wrap an R function in a Python function with the same signature.}
\usage{
ee_pyfunc(f)
}
\arguments{
\item{f}{An R function}
}
\value{
A Python function that calls the R function \code{f} with the same signature.
}
\description{
This function could wrap an R function in a Python function with
the same signature. Note that the signature of the R function
must not contain esoteric Python-incompatible constructs.
}
