% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_remove.R
\name{ee_remove_credentials}
\alias{ee_remove_credentials}
\title{Delete Credentials}
\usage{
ee_remove_credentials(email = "not_defined", quiet = FALSE)
}
\arguments{
\item{email}{Character (optional, e.g. \code{pinkiepie@gmail.com}).
The directory (all user credentials) to delete.}

\item{quiet}{Logical (optional). Suppress info messages.}
}
\description{
Delete all the credentials according to a specific user. The credentials
are saved in the path \code{rgee::ee_get_earthengine_path()} after running
successfully at least once  \code{ee_Initialize(...)}. However, if you run
\code{ee_Initialize(...)} with the same email argument the new credentials
will be overwritten.
}
\examples{
{
\dontrun{
ee_remove_credentials()
ee_remove_credentials('pinkiepie.us')
ee_remove_credentials('`pinkiepie.us@gmail.com`')
}
}
}
