% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_upload.R
\name{ee_upload}
\alias{ee_upload}
\alias{ee_upload.character}
\alias{ee_upload.sf}
\alias{ee_upload.stars}
\alias{ee_upload.stars_proxy}
\title{Upload sf or stars objects to GEE asset}
\usage{
ee_upload(x, ...)

\method{ee_upload}{character}(x, ..., filename, bucket = NULL,
  properties = getOption("rgee.upload.properties"),
  selenium_params = getOption("rgee.selenium.params"), quiet = FALSE)

\method{ee_upload}{sf}(x, ..., filename, bucket = NULL,
  selenium_params = getOption("rgee.selenium.params"), quiet = FALSE)

\method{ee_upload}{stars}(x, ..., filename, bucket = NULL,
  properties = getOption("rgee.upload.properties"),
  selenium_params = getOption("rgee.selenium.params"), quiet = FALSE)

\method{ee_upload}{stars_proxy}(x, ..., filename, bucket = NULL,
  properties = getOption("rgee.upload.properties"),
  selenium_params = getOption("rgee.selenium.params"), quiet = FALSE)
}
\arguments{
\item{x}{character, sf or stars object.}

\item{...}{ignored}

\item{filename}{character. images destination path, full path for upload to Google Earth Engine,
e.g. users/pinkiepie/myponycollection.}

\item{bucket}{bucketname you are uploading to}

\item{properties}{data.frame; \href{https://developers.google.com/earth-engine/image_manifest}{Manifest upload}.}

\item{selenium_params}{character; determines how each pixel at a given level of the pyramid is
computed from the aggregation. "MEAN" (default), "MODE" and "SAMPLE".}

\item{quiet}{logical; suppress info message}
}
\description{
Send either images or vectors to Google Earth Engine asset using Selenium
}
\details{
TALKING ABOUT SELENIUM
}
