% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_upload.R
\name{ee_upload}
\alias{ee_upload}
\alias{ee_upload.character}
\alias{ee_upload.sf}
\alias{ee_upload.stars}
\alias{ee_upload.stars_proxy}
\title{Upload sf or stars objects into a GEE asset}
\usage{
ee_upload(x, ...)

\method{ee_upload}{character}(x, ..., filename, bucket = NULL,
  properties = getOption("rgee.upload.properties"),
  selenium_params = getOption("rgee.selenium.params"), quiet = FALSE)

\method{ee_upload}{sf}(x, ..., filename, bucket = NULL,
  selenium_params = getOption("rgee.selenium.params"), quiet = FALSE)

\method{ee_upload}{stars}(x, ..., filename, bucket = NULL,
  properties = getOption("rgee.upload.properties"),
  selenium_params = getOption("rgee.selenium.params"), quiet = FALSE)

\method{ee_upload}{stars_proxy}(x, ..., filename, bucket = NULL,
  properties = getOption("rgee.upload.properties"),
  selenium_params = getOption("rgee.selenium.params"), quiet = FALSE)
}
\arguments{
\item{x}{character, sf or stars object.}

\item{...}{ignored}

\item{filename}{character. images destination path, full path for upload to Google Earth Engine,
e.g. users/pinkiepie/myponycollection.}

\item{bucket}{bucketname you are uploading to}

\item{properties}{data.frame; \href{https://developers.google.com/earth-engine/image_manifest}{Manifest upload}.}

\item{selenium_params}{character; determines how each pixel at a given level of the pyramid is
computed from the aggregation. "MEAN" (default), "MODE" and "SAMPLE".}

\item{quiet}{logical; suppress info message}
}
\description{
Upload either images or vectors to Google Earth Engine asset
}
\details{
TALKING ABOUT SELENIUM
}
\examples{
library(rgee)
library(stars)
library(sf)
ee_Initialize(user_gmail = 'csaybar@gmail.com')

filename <- "users/csaybar/rgee_upload/"
ee_manage_create(filename)

tif = system.file("tif/geomatrix.tif", package = "stars")
geomatrix = read_stars(tif) \%>\% st_warp(crs=st_crs(4326))
delta_geomatrix <- c(attr(geomatrix,'dimensions')$x$delta,attr(geomatrix,'dimensions')$y$delta*-1)
plot(geomatrix)

ee_upload(x = geomatrix,filename = paste0(filename,"geomatrix"))
ee_geomatrix <- ee$Image(paste0(filename,"geomatrix"))
geomatrix_stars <- ee_as_thumbnail(x = ee_geomatrix,
                                   scale = delta_geomatrix,
                                   vizparams = list(min = 0, max = 255))
geomatrix_stars[geomatrix_stars<=0]=NA
names(geomatrix_stars) <- 'geomatrix.tif'
plot(geomatrix_stars)
\dontrun{
nc <- st_read(system.file("shp/arequipa.shp", package="rgee"))
nc_s <- suppressWarnings(st_simplify(nc, preserveTopology = TRUE, dTolerance = 0.05))
ee_upload(x = nc_s,filename = paste0(filename,"arequipa"))
ee_monitoring()
help(ee_upload)
}
}
