% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_upload_toasset.R
\name{ee_upload_toasset}
\alias{ee_upload_toasset}
\alias{ee_upload_toasset.character}
\alias{ee_upload_toasset.sf}
\alias{ee_upload_toasset.stars}
\alias{ee_upload_toasset.stars_proxy}
\title{Upload sf or stars objects to GEE asset}
\usage{
ee_upload_toasset(x, ...)

\method{ee_upload_toasset}{character}(x, ..., gmail_account,
  gee_asset_path, properties = NULL, pyramiding_policy = "MEAN",
  one_by_one = TRUE, quiet = FALSE)

\method{ee_upload_toasset}{sf}(x, ..., gmail_account, asset_filename,
  gee_asset_path, quiet = FALSE)

\method{ee_upload_toasset}{stars}(x, ..., gmail_account, gee_asset_path,
  type = "Float32", properties = NULL, pyramiding_policy = "MEAN",
  one_by_one = TRUE, quiet = FALSE)

\method{ee_upload_toasset}{stars_proxy}(x, ..., gmail_account,
  gee_asset_path, properties = NULL, pyramiding_policy = "MEAN",
  one_by_one = TRUE, quiet = FALSE)
}
\arguments{
\item{x}{sf or stars object.}

\item{...}{ignored}

\item{gmail_account}{character; google account name (gmail address).}

\item{gee_asset_path}{character; images destination path, full path for upload to Google Earth Engine,
e.g. users/pinkiepie/myponycollection.}

\item{properties}{data.frame; \href{https://developers.google.com/earth-engine/image_manifest}{Manifest upload}.
See details}

\item{pyramiding_policy}{character; determines how each pixel at a given level of the pyramid is
computed from the aggregation. "MEAN" (default), "MODE" and "SAMPLE".}

\item{one_by_one}{TODO}

\item{quiet}{logical; suppress info message}

\item{asset_filename}{asset filename}

\item{type}{TODO}
}
\description{
Send either images or vectors to Google Earth Engine asset through Selenium
}
\examples{
\dontrun{
library(rgee)
}
}
