% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Initialize.R
\name{ee_user_info}
\alias{ee_user_info}
\title{Display credentials info of initialized user}
\usage{
ee_user_info()
}
\description{
Display credentials info of initialized user
a table.
}
\examples{
library(rgee)
ee_Initialize()
ee_user_info()
}
