% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Initialize.R
\name{ee_user_info}
\alias{ee_user_info}
\title{Display credentials and general info of the initialized user}
\usage{
ee_user_info()
}
\description{
Display credentials and general info of the initialized user
}
\examples{
\dontrun{
library(rgee)
ee_reattach() # reattach ee as a reserved word
ee_Initialize()
ee_user_info()
}
}
