% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_utils.R
\name{ee_utils_py_to_r}
\alias{ee_utils_py_to_r}
\title{Convert between Python and R objects}
\usage{
ee_utils_py_to_r(x)
}
\arguments{
\item{x}{A python object}
}
\value{
An R object
}
\description{
Convert between Python and R objects
}
\seealso{
Other ee_utils functions: 
\code{\link{ee_utils_pyfunc}()},
\code{\link{ee_utils_shp_to_zip}()}
}
\concept{ee_utils functions}
