% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Date.R
\name{eedate_to_rdate}
\alias{eedate_to_rdate}
\title{Pass Earth Engine date object to R}
\usage{
eedate_to_rdate(ee_date, js = FALSE)
}
\arguments{
\item{ee_date}{EE date object (ee$Date)}

\item{js}{Logical. If TRUE, return a numeric
date otherwise return the date as a POSIXct object.}
}
\description{
Function to pass objects of class ee$Date
to as.POSIXct or numeric (unix time stamp)
}
\examples{
library(rgee)
ee_reattach()
ee_Initialize()
eeDate <- ee$Date$fromYMD(1980,1,1)
eedate_to_rdate(eeDate)
}
