% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-upload.R
\name{gcs_to_ee_table}
\alias{gcs_to_ee_table}
\title{Move a zipped shapefile from GCS to their EE Assets}
\usage{
gcs_to_ee_table(
  manifest,
  command_line_tool_path = NULL,
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{manifest}{Character. manifest upload file. See \code{\link{ee_utils_create_manifest_table}}.}

\item{command_line_tool_path}{Character. Path to the Earth Engine command line
tool (CLT). If NULL, rgee assumes that CLT is set in the system PATH.
(ignore if \code{via} is not defined as "gcs_to_asset").}

\item{overwrite}{Logical. If TRUE, the assetId will be overwritten if
it exists.}

\item{quiet}{Logical. Suppress info message.}
}
\value{
Character. The Earth Engine asset ID.
}
\description{
Move a zipped shapefile from GCS to their EE Assets
}
\examples{
\dontrun{
library(rgee)
library(sf)
ee_Initialize(gcs = TRUE)

# Create sf object
nc <- st_read(system.file("shape/nc.shp", package="sf"))
assetId <- sprintf("\%s/\%s",ee_get_assethome(),'sf_nc')

# Method 1
# 1. Pass the sf to a zip file
zipfile <- ee_utils_shp_to_zip(nc)

# # 2. From local to gcs
# gs_uri <- local_to_gcs(x = zipfile, bucket = 'rgee_dev')
#
# # 3. Pass the sf to a zip file
# gcs_to_ee_table(
#   gs_uri = gs_uri,
#   overwrite = TRUE,
#   assetId = assetId
# )
#
# # OPTIONAL: Monitoring progress
# ee_monitoring()
#
# # OPTIONAL: Display results
# ee_sf_01 <- ee$FeatureCollection(assetId)
# Map$centerObject(ee_sf_01)
# Map$addLayer(ee_sf_01)
#
# # Method 2
# ee_sf_02 <- sf_as_ee(x = nc,
#                      assetId = assetId,
#                      overwrite = TRUE,
#                      bucket = "rgee_dev",
#                      via = 'gcs')
# Map$centerObject(ee_sf_02)
# Map$addLayer(ee_sf_02)
}
}
