% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_Date.R
\name{rdate_to_eedate}
\alias{rdate_to_eedate}
\title{Pass an R date object to Earth Engine}
\usage{
rdate_to_eedate(date, timestamp = FALSE)
}
\arguments{
\item{date}{R date object}

\item{timestamp}{Logical. If TRUE, return the date in milliseconds
from the Unix Epoch (1970-01-01 00:00:00 UTC) otherwise return a
EE date object. By default FALSE.}
}
\value{
\code{rdate_to_eedate} will return a numeric timestamp or
an EE Date depending on the \code{timestamp} argument.
}
\description{
Pass an R date object ("Date", "Numeric", "character", "POSIXt",
and "POSIXct") to Google Earth Engine (ee$Date).
}
\examples{
library(rgee)
ee_reattach()
ee_Initialize()
rdate_to_eedate('2000-01-01')
rdate_to_eedate(315532800000) # float number
}
