% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgee-package.R
\docType{package}
\name{rgee-package}
\alias{rgee}
\alias{rgee-package}
\title{rgee: R bindings for calling the Earth Engine API}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides bindings to the Google Earth Engine Python API and set tools for automating processes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/csaybar/rgee}
  \item Report bugs at \url{https://github.com/csaybar/rgee/issues}
}

}
\author{
\strong{Maintainer}: Cesar Aybar \email{csaybar@gmail.com} (\href{https://orcid.org/0000-0003-2745-9535}{ORCID})

Other contributors:
\itemize{
  \item Wu Qiusheng \email{giswqs@gmail.com} (\href{https://orcid.org/0000-0001-5437-4073}{ORCID}) [contributor]
  \item Roy Yali \email{ryali93@gmail.com} (\href{https://orcid.org/0000-0003-4542-3755}{ORCID}) [contributor]
  \item Antony Barja \email{antony.barja8@gmail.com} (\href{https://orcid.org/0000-0001-5921-2858}{ORCID}) [contributor]
  \item Gabriel Carrasco \email{gabriel.carrasco@upch.pe} (\href{https://orcid.org/0000-0002-6945-0419}{ORCID}) [contributor]
  \item MariaElena Adauto \email{2a.mariaelena@gmail.com} (\href{https://orcid.org/0000-0002-2154-2429}{ORCID}) [contributor]
}

}
\keyword{internal}
