% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgee-package.R
\docType{package}
\name{rgee-package}
\alias{rgee}
\alias{rgee-package}
\title{rgee: R bindings for calling the Earth Engine API}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

R bindings for calling the Earth Engine API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/csaybar/rgee}
  \item Report bugs at \url{https://github.com/csaybar/rgee/issues}
}

}
\author{
\strong{Maintainer}: Cesar Aybar \email{csaybar@gmail.com} (0000-0003-2745-9535)

Other contributors:
\itemize{
  \item Julio Contreras \email{julius013199@gmail.com} (0000-0001-5579-8919) [contributor]
  \item Roy Yali \email{ryali93@gmail.com} (0000-0003-4542-3755) [contributor]
  \item MariaElena Aduato \email{2a.mariaelena@gmail.com} (0000-0002-2154-2429) [contributor]
}

}
\keyword{internal}
