% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_operators.R
\name{slider}
\alias{slider}
\alias{|,EarthEngineMap,EarthEngineMap-method}
\title{EarthEngineMap | EarthEngineMap provides a slider in the middle to compare two maps.}
\usage{
\S4method{|}{EarthEngineMap,EarthEngineMap}(e1, e2)
}
\arguments{
\item{e1}{a leaflet or mapview map, or NULL.}

\item{e2}{a leaflet or mapview map, or NULL.}
}
\description{
EarthEngineMap | EarthEngineMap provides a slider in the middle to compare two maps.
}
\author{
tim-salabim. Adapted from mapview code.
}
